Augment data with information from a(n) iZIP model object

Augment accepts a model object and a dataset and adds information about each observation in the dataset. Most commonly, this includes predicted values in the .fitted column, residuals in the .resid column, and standard errors for the fitted values in a .se.fit column. New columns always begin with a . prefix to avoid overwriting columns in the original dataset.

# S3 method for izip
augment(
  x,
  data = model.frame.izip(x),
  newdata = NULL,
  type.predict = c("link", "response"),
  type.residuals = c("deviance", "pearson", "response"),
  se_fit = FALSE,
  ...
)

Arguments

x
an object class 'izip' object, obtained from a call to glm.izip

data
A base::data.frame or tibble::tibble() containing the original data that was used to produce the object x. Defaults to model.frame.izip(x) so that augment(my_fit) returns the augmented original data. Do not pass new data to the data argument. Augment will report information such as influence and cooks distance for data passed to the data argument. These measures are only defined for the original training data.

newdata
A base::data.frame() or tibble::tibble() containing all the original predictors used to create x. Defaults to NULL, indicating that nothing has been passed to newdata. If newdata is specified, the data argument will be ignored.

type.predict
Passed to predict.izip() type argument. Defaults to "link".

type.residuals
Passed to residuals.izip() type arguments. Defaults to "deviance".

se_fit
Logical indicating whether or not a .se.fit column should be added to the augmented output. Defaults to FALSE.

...
Additional arguments. Not used. Needed to match generic signature only. Cautionary note: Misspelled arguments will be absorbed in ..., where they will be ignored. If the misspelled argument has a default value, the default value will be used. For example, if you pass conf.level = 0.9, all computation will proceed using conf.level = 0.95. Additionally, if you pass newdata = my_tibble to an augment() method that does not accept a newdata argument, it will use the default value for the data argument.

Value

A tibble::tibble() with columns:

.cooksd

Cooks distance.

.fitted

Fitted or predicted value.

.hat

Diagonal of the hat matrix.

.resid

The difference between observed and fitted values.

.se.fit

Standard errors of fitted values.

.sigma

Estimated residual standard deviation when corresponding observation is dropped from model.

.std.resid

Standardised residuals.

Details

Users may pass data to augment via either the data argument or the newdata argument. If the user passes data to the data argument, it must be exactly the data that was used to fit the model object. Pass datasets to newdata to augment data that was not used during model fitting. This still requires that all columns used to fit the model are present.

Augment will often behave differently depending on whether data or newdata is given. This is because there is often information associated with training observations (such as influences or related) measures that is not meaningfully defined for new observations.

For convenience, many augment methods provide default data arguments, so that augment(fit) will return the augmented training data. In these cases, augment tries to reconstruct the original data based on the model object with varying degrees of success.

The augmented dataset is always returned as a tibble::tibble with the same number of rows as the passed dataset. This means that the passed data must be coercible to a tibble.

We are in the process of defining behaviours for models fit with various na.action arguments, but make no guarantees about behaviour when data is missing at this time.

Examples

data(bioChemists)
M_bioChem <- glm.izip(art ~ ., data = bioChemists)
augment(M_bioChem)
#> # A tibble: 915 x 11
#>      art fem   mar    kid5   phd  ment .fitted .resid .std.resid    .hat .cooksd
#>    <int> <fct> <fct> <dbl> <dbl> <int>   <dbl>  <dbl>      <dbl>   <dbl>   <dbl>
#>  1     0 Men   Marr…     0  2.52     7  0.682   -1.83      -1.83 0.00481 1.22e-3
#>  2     0 Women Sing…     0  2.05     6  0.265   -1.49      -1.49 0.00457 7.27e-4
#>  3     0 Women Sing…     0  3.75     6  0.282   -1.50      -1.50 0.00368 5.95e-4
#>  4     0 Men   Marr…     1  1.18     3  0.380   -1.57      -1.58 0.00572 1.03e-3
#>  5     0 Women Sing…     0  3.75    26  0.776   -1.91      -1.92 0.00788 2.25e-3
#>  6     0 Women Marr…     2  3.59     2 -0.0386  -1.29      -1.29 0.00613 7.17e-4
#>  7     0 Women Sing…     0  3.19     3  0.202   -1.44      -1.44 0.00330 4.90e-4
#>  8     0 Men   Marr…     2  2.96     4  0.233   -1.46      -1.47 0.00526 8.10e-4
#>  9     0 Men   Sing…     0  4.62     6  0.519   -1.68      -1.69 0.00721 1.52e-3
#> 10     0 Women Marr…     0  1.25     0  0.267   -1.49      -1.49 0.00598 9.57e-4
#> # … with 905 more rows