bioChemists data set

This data set gives the number of article produced by a sample of 915 graduate students during the last 3 years of their biochemistry PhD program, along with some information on the graduates like gender, marital status (mar), the number of kids under 5 (kid5), how prestigious the PhD department is (phd) and the mentor’s publication record over the same period (ment). The bioChemist data has 915 observations.

data(bioChemists)

Format

A data frame with 915 observations on 6 variables.

art

count of articles produced during last 3 years of Ph.D.

fem

factor indicating gender of student, with levels Men and Women

mar

factor indicating marital status of student, with levels Single and Married

kid5

number of children aged 5 or younger

phd

prestige of Ph.D. department

ment

count of articles produced by Ph.D. mentor during last 3 years

Source

The pscl package

References

Jackman, S. (2020). pscl: Classes and Methods for R Developed in the Political Science Computational Laboratory. United States Studies Centre, University of Sydney. Sydney, New South Wales, Australia. R package version 1.5.5. URL https://github.com/atahk/pscl/

Long, J. Scott. 1990. The origins of sex differences in science. Social Forces. 68(3):1297-1316.

Long, J. Scott. 1997. Regression Models for Categorical and Limited Dependent Variables. Thousand Oaks, California: Sage.

Examples

## For examples see example(glm.izip)