Confidence Intervals for iZIP Model Parameters

Computes confidence intervals for one or more parameters in a fitted model.

# S3 method for izip
confint(object, parm, level = 0.95, ...)

# S3 method for tsizip
confint(object, parm, level = 0.95, ...)

Arguments

object
an object class 'izip' or 'tsizip', obtained from a call to glm.izip or tsglm.izip.

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names (comparing to those provided by coef()) . If missing, all parameters are considered.

level
the confidence level required.

...
other arguments passed to or from other methods (currently unused).

Value

A matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in % (by default 2.5% and 97.5%).

Examples

data(bioChemists)
M_bioChem <- glm.izip(art ~ ., data = bioChemists)
confint(M_bioChem)

#>                   2.5 %      97.5 %
#> (Intercept)  0.09261810  0.55651638
#> femWomen    -0.35207646 -0.10516958
#> marMarried   0.02051290  0.29726928
#> kid5        -0.28028826 -0.09879790
#> phd         -0.04917707  0.06941332
#> ment         0.02042481  0.02894963

confint(M_bioChem, param = "ment", level = 0.9)

#>                     5 %        95 %
#> (Intercept)  0.12990939  0.51922509
#> femWomen    -0.33222841 -0.12501762
#> marMarried   0.04276045  0.27502173
#> kid5        -0.26569884 -0.11338733
#> phd         -0.03964397  0.05988022
#> ment         0.02111009  0.02826435

data(arson)
M_arson <- tsglm.izip(arson ~ 1, past_mean_lags = 1, past_obs_lags = c(1, 2))
confint(M_arson)

#>                    2.5 %    97.5 %
#> (Intercept) -0.109465665 1.7724965
#> beta_1      -0.137404909 0.1901232
#> beta_2       0.001429144 0.3497717
#> alpha_1     -0.913769353 0.9137694