Fit an interpretable Zero-Inflated Poisson Generalized Linear Model

The function glm.izip is used to fit an interpretable zero-inflated Poisson generalized linear model with a log-link.

glm.izip(
  formula,
  data,
  ref.lambda = NULL,
  offset = NULL,
  subset,
  contrasts = NULL,
  na.action
)

Arguments

formula
an object of class 'formula': a symbolic description of the model to be fitted to the mean via log-link.

data
an optional data frame containing the variables in the model

ref.lambda
the rate of a Poisson distribution that baseline zero-inflated odds based on.

offset
this can be used to specify an a priori known component to be included

subset
an optional vector specifying a subset of observations to be used in the fitting process.

na.action
a function which indicates what should happen when the data contain NAs. The default is set by the na.action setting of options, and is na.fail if that is unset. The ‘factory-fresh’ default is na.omit. Another possible value is NULL, no action. Value na.exclude can be useful.

contrast
optional lists. See the contrasts.arg of model.matrix.default.

Value

A fitted model object of class izip similar to one obtained from glm or glm.nb.

The function summary (i.e., summary.izip) can be used to obtain and print a summary of the results.

The functions plot (i.e., plot.izip) and autoplot can be used to produce a range of diagnostic plots.

The generic assessor functions coef (i.e., coef.izip), logLik (i.e., logLik.izip) fitted (i.e., fitted.izip), nobs (i.e., nobs.izip), AIC (i.e., AIC.izip) and residuals (i.e., residuals.izip) can be used to extract various useful features of the value returned by glm.izip.

An object class 'glm.izip' is a list containing at least the following components:

coefficients

a named vector of coefficients

stderr

approximate standard errors (using observed rather than expected information) for mean coefficients

residuals

the response residuals (i.e., observed-fitted)

fitted_values

the fitted mean values

rank

the numeric rank of the fitted linear model for mean

linear_predictors

the linear fit for mean on log scale

df_residuals

the residuals degrees of freedom

df_null

the residual degrees of freedom for the null model

null_deviance

The deviance for the null model. The null model will include only the intercept.

deviance; residual_deviance

The residual deviance of the model

y

the y vector used.

X

the model matrix for mean

model

the model frame for regression

call

the matched call

formula

the formula supplied for regression

terms

the terms object used for regression

data

the data argument

offset

the offset vector used

Details

Fit an interpretable zero-inflated Poisson regression using maximum likelihood estimation.

The model is

$$Y_i ~ ZIP_{\nu}(\mu_i | \lambda = \lambda_{ref}),$$

where

$$E(Y_i) = \mu_i = exp(x_i^T \beta),$$

\(x_i\) are some covariates.

\(\nu \ge 0\) is the baseline zero-inflated odds relative to a Poisson with rate \(\lambda_{ref}\).

References

Huang, A. and Fung, T. (2020). Zero-inflated Poisson exponential families, with applications to time-series modelling of counts.

Examples

data()