Non-randomized Probability Integral Transform

Functions to produce the non-randomized probability integral transform (PIT) to check the adequacy of the distributional assumption of the iZIP model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

PredProb.izip(object)

PIT.izip(object, bins = 10)

Arguments

object
an object class "izip", obtained from a call to glm.izip.

bins
numeric; the number of bins shown in the PIT histogram or the PIT Q-Q plot.

Value

compPredprob returns a list with values:

upper

the predictive cumulative probabilities used as the upper bound for computing the non-randomized PIT.

lower

the predictive cumulative probabilities used as the upper bound for computing the non-randomized PIT.

izipPIT returns a list with values:
conditionalPIT

the conditional probability integral transformation given the observed counts.

PIT

the probability integral transformation.

Details

These functions are used to obtain the predictive probabilities and the probability integral transform for a fitted COM-Poisson model. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

References

Czado, C., Gneiting, T. and Held, L. (2009). Predictive model assessment for count data. Biometrics, 65, 1254--1261.

Dunsmuir, W.T.M. and Scott, D.J. (2015). The glarma Package for Observation-Driven Time Series Regression of Counts. Journal of Statistical Software, 67, 1--36.

Examples

data(takeoverbids)

#> Warning: data set ‘takeoverbids’ not found

M.bids <- glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght
    + bidprem + insthold + size + sizesq + regulatn, data=takeoverbids)

#> Error in glm.cmp(numbids ~ leglrest + rearest + finrest + whtknght + bidprem +     insthold + size + sizesq + regulatn, data = takeoverbids): could not find function "glm.cmp"

izipPredProb(M.bids)

#> Error in izipPredProb(M.bids): could not find function "izipPredProb"

izipPIT(M.bids)

#> Error in izipPIT(M.bids): could not find function "izipPIT"