Extract iZIP Model Residuals

residuals is a generic function which extracts model residuals from objects returned by the modelling function glm.izip or tsglm.izip. resid is an alias for residuals .

# S3 method for izip
residuals(object, type = c("deviance", "pearson", "response"), ...)

# S3 method for tsizip
residuals(object, type = c("response", "pearson"), ...)

Arguments

object
an object class 'izip' or 'tsizip', obtained from a call to glm.izip or tsglm.izip.

type
the type of residuals which should be returned. The alternatives are: 'deviance' (default), 'pearson' and 'response' for an 'izip' object; 'response' (default), and 'pearson' for an 'tsizip' object. Can be abbreviated.

...
other arguments passed to or from other methods (currently unused).

Value

Residuals extracted from the object object.

See also

Examples

data(bioChemists)
M_bioChem <- glm.izip(art ~ ., data = bioChemists)
# deviance residuals
residuals(M_bioChem)

#>            1            2            3            4            5            6 
#> -1.825642155 -1.486723228 -1.499133994 -1.572134000 -1.914640906 -1.286031924 
#>            7            8            9           10           11           12 
#> -1.442602066 -1.464134989 -1.683610523 -1.488644178 -1.842628890 -1.606780056 
#>           13           14           15           16           17           18 
#> -1.424979517 -1.578335984 -1.492579316 -1.408142134 -1.625293819 -1.642015862 
#>           19           20           21           22           23           24 
#> -1.514260389 -1.558115091 -1.607810058 -1.321539322 -1.635605860 -1.604487817 
#>           25           26           27           28           29           30 
#> -1.289691019 -1.545085650 -1.511224334 -1.596484138 -1.391601606 -1.735821093 
#>           31           32           33           34           35           36 
#> -1.386896967 -1.865327806 -1.612765243 -1.546242064 -1.519069285 -1.384283623 
#>           37           38           39           40           41           42 
#> -1.382611334 -1.743762367 -1.553644892 -1.600261929 -1.743762367 -1.419250705 
#>           43           44           45           46           47           48 
#> -1.657411408 -1.422698635 -1.662544510 -1.516538701 -1.695318710 -2.308052317 
#>           49           50           51           52           53           54 
#> -1.657276787 -1.660952748 -1.662545341 -1.246626354 -1.611550533 -1.865327806 
#>           55           56           57           58           59           60 
#> -1.561625363 -1.579082090 -1.425737372 -1.549749790 -1.555636243 -1.552800674 
#>           61           62           63           64           65           66 
#> -1.492621598 -1.523397604 -1.463828444 -1.541898074 -1.516074293 -1.584449122 
#>           67           68           69           70           71           72 
#> -1.915140211 -1.795444158 -1.524145330 -1.377641437 -1.584053838 -1.250268000 
#>           73           74           75           76           77           78 
#> -1.510511568 -1.601150351 -1.561913498 -1.478922922 -1.390679117 -1.542599362 
#>           79           80           81           82           83           84 
#> -1.619193166 -1.477458251 -1.418077237 -1.390662508 -1.911651899 -1.800281916 
#>           85           86           87           88           89           90 
#> -2.121683579 -1.399050566 -2.153572426 -1.500844067 -2.145152623 -1.681425393 
#>           91           92           93           94           95           96 
#> -1.558474203 -1.499678483 -1.727974363 -2.121683579 -1.659230929 -1.816369232 
#>           97           98           99          100          101          102 
#> -1.725332781 -1.516191688 -1.317526985 -1.865327806 -1.481359013 -1.523397604 
#>          103          104          105          106          107          108 
#> -1.505921852 -1.576700108 -1.405094882 -1.568493388 -1.639494928 -1.486723228 
#>          109          110          111          112          113          114 
#> -1.316782864 -1.284614514 -1.719080238 -1.385087134 -1.405963426 -1.781117729 
#>          115          116          117          118          119          120 
#> -1.458366973 -1.598460993 -1.751644839 -1.367798901 -1.579082090 -1.570971542 
#>          121          122          123          124          125          126 
#> -1.523005458 -1.550995397 -1.719706495 -1.401403951 -1.416359166 -1.791133656 
#>          127          128          129          130          131          132 
#> -1.535063949 -1.722335084 -1.546242064 -1.584977256 -1.420970119 -2.036202660 
#>          133          134          135          136          137          138 
#> -1.587251763 -1.446188934 -1.541562181 -2.032014580 -1.624779529 -1.462687271 
#>          139          140          141          142          143          144 
#> -1.290309747 -1.430834987 -1.367468754 -1.513401938 -1.549288501 -1.422698635 
#>          145          146          147          148          149          150 
#> -1.417389448 -1.330900989 -1.671401614 -1.417389448 -1.723634659 -1.389443531 
#>          151          152          153          154          155          156 
#> -1.584210415 -1.741372307 -1.562266086 -1.463608937 -1.590645895 -1.618791414 
#>          157          158          159          160          161          162 
#> -1.486963402 -1.510437560 -1.584865968 -1.851728971 -1.553167076 -1.412654754 
#>          163          164          165          166          167          168 
#> -1.548786096 -1.541410588 -1.732062359 -1.479840630 -1.683977930 -1.604626386 
#>          169          170          171          172          173          174 
#> -1.678600922 -1.639005569 -1.543035875 -1.529147685 -1.404258084 -1.557423633 
#>          175          176          177          178          179          180 
#> -1.425342551 -1.429064519 -1.650763512 -1.485271746 -1.353985906 -1.523397604 
#>          181          182          183          184          185          186 
#> -1.367732864 -1.924010069 -1.724638692 -1.762347863 -1.643759642 -1.653072482 
#>          187          188          189          190          191          192 
#> -1.456783918 -1.615528194 -1.375978211 -1.528996778 -1.382878746 -1.481000848 
#>          193          194          195          196          197          198 
#> -1.387635072 -1.692263973 -1.634761631 -1.663654229 -1.501147123 -1.253921202 
#>          199          200          201          202          203          204 
#> -1.867107400 -1.242372505 -1.621012830 -1.620238283 -1.436464176 -1.458203892 
#>          205          206          207          208          209          210 
#> -1.531175742 -1.262167715 -1.553396521 -1.563008543 -1.384546279 -1.496965918 
#>          211          212          213          214          215          216 
#> -1.415555295 -1.534457024 -1.511594642 -1.423821748 -1.523902507 -1.743762367 
#>          217          218          219          220          221          222 
#> -1.750231073 -1.819423783 -1.607173516 -1.804184587 -1.737483189 -1.609182012 
#>          223          224          225          226          227          228 
#> -1.653020745 -1.561683333 -1.541367440 -1.682487843 -1.563008543 -1.413819765 
#>          229          230          231          232          233          234 
#> -1.605739813 -1.757313916 -1.517905431 -2.124556953 -1.556705615 -1.713615662 
#>          235          236          237          238          239          240 
#> -1.719706495 -1.631570184 -1.443795410 -1.880423497 -1.523397604 -1.660472898 
#>          241          242          243          244          245          246 
#> -1.399708874 -1.406368947 -1.641555988 -1.634609232 -1.601150351 -1.552800674 
#>          247          248          249          250          251          252 
#> -1.486236954 -1.659674691 -1.412243150 -1.405891985 -1.977360273 -1.701497749 
#>          253          254          255          256          257          258 
#> -1.463400384 -1.756459561 -1.450343856 -1.432191800 -1.533207720 -1.591406597 
#>          259          260          261          262          263          264 
#> -1.633065746 -1.716422369 -1.615581693 -1.577343066 -1.523397604 -1.392275174 
#>          265          266          267          268          269          270 
#> -1.439050203 -1.731718809 -1.399322264 -2.431375219 -1.477046875 -1.639021729 
#>          271          272          273          274          275          276 
#> -1.498358140 -1.747849394 -1.520830808 -1.673239731 -1.691385109 -0.615430680 
#>          277          278          279          280          281          282 
#> -0.169356103 -0.673677217  0.057085403 -0.380558632 -0.263557658 -0.500045001 
#>          283          284          285          286          287          288 
#> -0.261545839 -0.251279927 -0.702169482 -0.153733172 -0.609189053 -0.733931677 
#>          289          290          291          292          293          294 
#> -0.460712991 -0.873698481 -0.715929014 -0.572215117 -0.554597735 -0.395737164 
#>          295          296          297          298          299          300 
#> -0.350691037 -0.523600700 -0.967607232  0.037432058 -0.432440418 -0.594750901 
#>          301          302          303          304          305          306 
#> -0.483722269 -0.445267159 -1.619931271 -0.347771299 -0.444861969 -0.504473058 
#>          307          308          309          310          311          312 
#> -0.502654337 -0.203924667 -0.670121988 -0.319503521 -0.124042845 -0.769075445 
#>          313          314          315          316          317          318 
#> -0.843171561 -0.712731579 -0.739386322 -0.258619462 -0.863401178 -0.277007879 
#>          319          320          321          322          323          324 
#> -0.300332150 -0.437053910 -0.578323287 -0.736039535 -0.203826075 -0.653702167 
#>          325          326          327          328          329          330 
#> -0.691452156 -0.387847174 -0.180497119 -3.376917704 -0.646443685 -0.398844541 
#>          331          332          333          334          335          336 
#> -0.512527377 -0.249874709 -0.756913006 -0.693332298 -0.322468068 -0.316823799 
#>          337          338          339          340          341          342 
#> -0.317132645 -0.348631948 -0.403527561 -0.249874709 -0.193004630 -0.261843344 
#>          343          344          345          346          347          348 
#> -0.364840953 -0.302832026 -0.207012009 -0.327555286 -0.403393703 -0.765888382 
#>          349          350          351          352          353          354 
#> -0.147935969 -0.190746418 -0.379886214 -0.528964675 -0.291680896 -0.315653053 
#>          355          356          357          358          359          360 
#> -0.503544519 -0.434559694 -0.171590265 -0.261784942 -0.225709496 -0.523823974 
#>          361          362          363          364          365          366 
#> -0.866080167 -0.581056419 -0.718916120 -0.598656836 -0.115830084 -0.386515924 
#>          367          368          369          370          371          372 
#> -0.099878913 -0.159377926 -0.171688626 -0.458422623 -0.191826703 -0.529034927 
#>          373          374          375          376          377          378 
#> -0.190580197 -0.394735991 -0.590168912 -0.349707269 -0.406293013 -0.612606106 
#>          379          380          381          382          383          384 
#> -0.555503010 -0.740728802 -1.073086159 -0.803766111 -0.281402775 -0.315622637 
#>          385          386          387          388          389          390 
#> -0.272956069 -0.594800342 -0.301501976 -0.788042588 -0.159349281 -0.085175278 
#>          391          392          393          394          395          396 
#> -0.631261568 -0.591316179 -0.214524718 -0.720557845 -0.612885950  0.053025325 
#>          397          398          399          400          401          402 
#> -0.155765209 -0.378593099 -0.483902387 -0.388422106 -1.185400778 -0.504624981 
#>          403          404          405          406          407          408 
#> -0.651862376 -0.334880696 -0.569053737 -0.336437654  0.043662557 -1.121417760 
#>          409          410          411          412          413          414 
#> -0.578394913 -0.295941201 -0.175587056 -0.471408901 -0.294511578 -0.706609053 
#>          415          416          417          418          419          420 
#> -0.111895503 -0.203924667 -0.350751481 -0.070863512 -0.582319481 -0.865267492 
#>          421          422          423          424          425          426 
#> -0.778406013 -0.005228627 -0.135832491 -0.203732284 -0.204607365 -0.552748401 
#>          427          428          429          430          431          432 
#> -0.754661038 -0.534893861 -0.524383361 -0.136794353 -0.442886613 -0.402675661 
#>          433          434          435          436          437          438 
#> -0.612885950 -0.791335015 -0.445944600 -0.595211262 -0.402351092 -0.646400209 
#>          439          440          441          442          443          444 
#> -0.495578173 -0.225635615 -0.702729557 -0.663863278 -0.297269416 -0.739266065 
#>          445          446          447          448          449          450 
#> -0.606650540 -0.508755428 -0.337993477 -0.430506436 -0.635879273 -0.220342567 
#>          451          452          453          454          455          456 
#> -0.813124187 -0.214792459 -0.135131686 -0.666739333 -1.806868057 -0.295527687 
#>          457          458          459          460          461          462 
#> -0.377179001 -0.548052266 -0.821260727 -0.276949136 -0.119204847 -0.251682560 
#>          463          464          465          466          467          468 
#> -0.288534319 -0.408425242 -0.551775480 -0.396821324 -0.623379554 -0.366835041 
#>          469          470          471          472          473          474 
#> -0.475343962 -0.427744155 -1.222332715 -0.582883909 -0.345084750 -0.258057777 
#>          475          476          477          478          479          480 
#> -0.350691037 -0.408110325 -0.442260268 -0.356841121 -0.171590265 -0.778406013 
#>          481          482          483          484          485          486 
#> -0.774246559 -0.774246559 -1.849383508 -0.514194032 -0.462537200 -0.504473058 
#>          487          488          489          490          491          492 
#> -0.560135050 -0.777813869 -0.437444355 -0.302525820 -0.469982156 -0.103672751 
#>          493          494          495          496          497          498 
#> -0.426635625 -0.833704534 -0.382486469 -0.496839461 -0.288695183 -0.425667490 
#>          499          500          501          502          503          504 
#> -0.575353939 -0.198709230 -0.618899122 -0.374195077 -0.359181971 -0.132832229 
#>          505          506          507          508          509          510 
#> -0.291820814 -0.500045001 -0.614576522 -1.065915709 -0.668915319 -0.340097446 
#>          511          512          513          514          515          516 
#> -0.522660857 -0.906091357 -0.545728286 -0.844261596 -0.189080261 -0.161094650 
#>          517          518          519          520          521          522 
#> -0.225970781 -0.245694216 -0.509574847 -0.489077540 -0.544152297  0.098744166 
#>          523          524          525          526          527          528 
#> -0.010848926  0.439536333  0.184180479  0.518980655  0.434889792  0.550455237 
#>          529          530          531          532          533          534 
#>  0.621362181  0.593177209  0.376128521  0.605327031  0.439042857  0.535647756 
#>          535          536          537          538          539          540 
#>  0.135766340  0.315080611  0.045164844  0.286785978  0.484891158  0.516511390 
#>          541          542          543          544          545          546 
#> -0.433713463  0.662348996  0.393141295 -0.881663040  0.116715072 -0.857857697 
#>          547          548          549          550          551          552 
#> -1.557895456 -0.061756213  0.114808168  0.423726682  0.522566302  0.622123685 
#>          553          554          555          556          557          558 
#>  0.397027285  0.203695388  0.218080445  0.663638873  0.100459122  0.507713822 
#>          559          560          561          562          563          564 
#>  0.343669618  0.137361085 -0.311242792  0.319776880  0.316035423  0.515005973 
#>          565          566          567          568          569          570 
#>  0.230535644  0.341036461  0.247766214  0.585785635 -0.542088183  0.052228133 
#>          571          572          573          574          575          576 
#>  0.258673298  0.348398642  0.274715563 -0.005613554  0.211227265  0.355597083 
#>          577          578          579          580          581          582 
#>  0.091066855  0.299330695  0.809647387  0.563207274  0.554119465  0.084878809 
#>          583          584          585          586          587          588 
#>  0.243907965  0.256293585  0.260829516  0.379992670  0.252567309  0.366906901 
#>          589          590          591          592          593          594 
#>  0.309983362 -0.559815631  0.364072999  0.624297390  0.281431925  0.364401206 
#>          595          596          597          598          599          600 
#>  0.013276448  0.418904751  0.666722358  0.152773226  0.513704464 -0.405976284 
#>          601          602          603          604          605          606 
#>  0.197508098  0.659022820  0.581088509  0.270971778  0.586588715 -0.611417561 
#>          607          608          609          610          611          612 
#>  0.218787125  0.483077815  0.076730748  0.333850723 -0.484898002  0.165514312 
#>          613          614          615          616          617          618 
#>  0.446029351  0.304963752  0.528833894  0.459920365  0.214385217 -0.443776440 
#>          619          620          621          622          623          624 
#>  0.240687296  0.159095431  0.023146995  0.251101504  0.097263750  0.509612147 
#>          625          626          627          628          629          630 
#> -0.148706038  0.078355830  0.820765774  0.199458868 -0.143834182 -0.125730211 
#>          631          632          633          634          635          636 
#>  0.054722730  0.586363528  0.525701567  0.527235040  0.657947600  0.105182293 
#>          637          638          639          640          641          642 
#> -0.238698406  0.502566872  0.565641303  0.045773003  0.471884588  0.063137489 
#>          643          644          645          646          647          648 
#> -0.349207657  0.356235985 -0.208460939  0.507854121  0.456607976  0.231632584 
#>          649          650          651          652          653          654 
#>  0.335590799  0.370531546  0.543782757 -0.125238606  0.521441519  0.462838141 
#>          655          656          657          658          659          660 
#>  0.638286487  0.156194563  0.421568082  0.176257024 -0.097068257  0.373200128 
#>          661          662          663          664          665          666 
#>  0.433595743  0.231510046  0.161179629  0.587132896  0.617220038 -0.208460939 
#>          667          668          669          670          671          672 
#>  0.263529456  0.603180808 -1.721390959  0.165593003  0.357687889  0.577337897 
#>          673          674          675          676          677          678 
#>  0.556005537  0.383572800  0.113789404  0.507038593 -0.045383427 -0.591440481 
#>          679          680          681          682          683          684 
#>  0.253931393  0.240687296 -0.510741706 -0.583726357  0.411542143  0.251863411 
#>          685          686          687          688          689          690 
#>  0.409716463  0.306304421  0.389532926  0.681753804  0.839323854  0.127297717 
#>          691          692          693          694          695          696 
#>  0.311089913  0.587792217 -0.021640039 -0.265565469  0.468128719  0.616353786 
#>          697          698          699          700          701          702 
#>  0.492854932  0.256697530  0.131474717  0.953300954  1.316499624  0.781136050 
#>          703          704          705          706          707          708 
#>  0.289683359  0.688164331  0.659380211  1.268142153  0.942980959  0.672246629 
#>          709          710          711          712          713          714 
#> -0.016480979  0.520979332  1.225400912  1.059425961 -0.613393651  0.587959524 
#>          715          716          717          718          719          720 
#>  0.672246629  0.732006325  1.155967616  1.008338194  0.925340432  1.015128959 
#>          721          722          723          724          725          726 
#>  0.668181658  0.524549337  1.300488458  0.938494388  0.511620245  0.487649906 
#>          727          728          729          730          731          732 
#>  0.912495073  1.042345527  0.811804183  0.622850336  0.946597907  1.241882345 
#>          733          734          735          736          737          738 
#>  0.984384567  0.590785644  0.827304510  0.126982466  1.113704080  1.215991804 
#>          739          740          741          742          743          744 
#>  1.160674207  0.792261068  0.437002803 -0.386733090  0.805026258 -0.024264334 
#>          745          746          747          748          749          750 
#>  0.457507338  1.192417888  0.451983207  0.777061539  1.132960237 -0.107084802 
#>          751          752          753          754          755          756 
#>  0.420587842  1.345258572  0.965619077  0.317609426  1.016081294  0.858316716 
#>          757          758          759          760          761          762 
#>  1.220998765  0.788621909  0.869180997  0.930019212  0.841103436  0.807977967 
#>          763          764          765          766          767          768 
#>  0.694754670  1.270850146  0.755093358  0.367020132  0.746502223  0.198878560 
#>          769          770          771          772          773          774 
#>  0.636586399  1.263905655 -0.656182196  0.662985819  1.263281258  0.980461089 
#>          775          776          777          778          779          780 
#>  0.580742673  0.675861225  0.194208541  1.211061514  1.101562211  0.936826568 
#>          781          782          783          784          785          786 
#>  1.447914539  1.165439912  1.295960129  1.250599799  1.163778064  0.575671689 
#>          787          788          789          790          791          792 
#>  1.383398481  1.096601429  1.441697963  1.848173024  1.510843214  1.902384688 
#>          793          794          795          796          797          798 
#>  1.334220584  1.672064225 -0.275924777  1.447467126  1.556288213  0.889374231 
#>          799          800          801          802          803          804 
#> -0.722406039  1.554415619  1.864243447  1.434312582 -0.832612326  1.324379046 
#>          805          806          807          808          809          810 
#>  1.448063413  1.276474016  1.395654599  1.085934764  1.511942903  0.984930045 
#>          811          812          813          814          815          816 
#>  1.636030694 -0.172154272  1.819305005  1.841500117  0.225604458  1.037941644 
#>          817          818          819          820          821          822 
#>  1.274414346  1.614774856  1.576482052  1.460107402  1.202971095  0.545512999 
#>          823          824          825          826          827          828 
#>  1.263361773  1.616829071  1.254294444  1.736864864  1.571295002  1.670199921 
#>          829          830          831          832          833          834 
#>  1.680992595  0.952199698  1.391758969  0.950479586  1.807714921  1.570172802 
#>          835          836          837          838          839          840 
#>  0.604653263  1.492246412  1.493367578  1.825726722  1.172679480  1.195263287 
#>          841          842          843          844          845          846 
#>  0.794778619  1.588657467  1.632541215  1.610232793  1.818681663  1.389020504 
#>          847          848          849          850          851          852 
#>  1.239189282  1.304529227  1.729897391  1.122535715  1.712894409  2.028858319 
#>          853          854          855          856          857          858 
#>  1.945799395  2.036683522  1.201486129  1.708683566  1.480782793  1.776486080 
#>          859          860          861          862          863          864 
#>  1.996465057  1.294378885  1.439248948  2.073698569  1.436583062  0.704391181 
#>          865          866          867          868          869          870 
#>  2.075615412  1.715328593  2.464063079  1.809890445  2.201566393  1.394638922 
#>          871          872          873          874          875          876 
#>  2.238411517 -0.086415973  1.900166396  1.585031187  1.845369006  2.280235386 
#>          877          878          879          880          881          882 
#>  2.363766123  2.402025246  1.353135956  1.940163071  2.075457154  1.843119246 
#>          883          884          885          886          887          888 
#>  2.601757561  2.240241640  1.973656227  2.146939100  2.548850176  2.110082667 
#>          889          890          891          892          893          894 
#>  2.674621115  2.440582149  2.447066310  2.712632837  2.541440921  2.775575322 
#>          895          896          897          898          899          900 
#>  2.857235390  2.927849885  2.355836724  3.201309605  3.148836531  2.542910895 
#>          901          902          903          904          905          906 
#>  1.289136622  2.540671598  2.381997736  2.382861489  2.101575476  2.393213728 
#>          907          908          909          910          911          912 
#>  2.909918211  3.122683023  1.921373512  3.800652493  4.918420253  3.614817563 
#>          913          914          915 
#>  5.089535998  5.349236232  4.953133262 

data(arson)
M_arson <- tsglm.izip(arson ~ 1, past_mean_lags = 1, past_obs_lags = c(1, 2))
# pearson residuals
resid(M_arson, type = "pearson")

#>            1            2            3            4            5            6 
#> -0.036550880  0.836222476 -0.051699480  1.450880387 -0.927200740 -1.040977484 
#>            7            8            9           10           11           12 
#>  0.164547828 -0.824529234  2.652438155  0.057940317  1.034677000 -0.927200740 
#>           13           14           15           16           17           18 
#>  0.491768886  0.109560269  0.640751137 -0.051699480 -0.169378115 -0.029287642 
#>           19           20           21           22           23           24 
#>  0.845623306  0.812346766  0.612993814  2.216502947  2.131178750 -1.152026059 
#>           25           26           27           28           29           30 
#> -0.386706862 -0.824529235 -0.892554580 -0.812086474  3.094450738  0.057940317 
#>           31           32           33           34           35           36 
#> -1.121313720 -0.892554581 -0.812086474  0.164547829 -0.824529234  0.879941788 
#>           37           38           39           40           41           42 
#> -0.836831742  0.669206027 -0.836831742 -0.968428480  0.164547828  0.136601053 
#>           43           44           45           46           47           48 
#> -0.904198589 -0.006306397  0.136601053  1.720534254  1.633671408 -1.072784473 
#>           49           50           51           52           53           54 
#>  0.491768885 -0.836831742 -0.149611226  1.097731339  0.812346767  1.414748381 
#>           55           56           57           58           59           60 
#> -1.001455274  0.491768886  0.109560269 -0.169378114 -0.904198589  0.879941787 
#>           61           62           63           64           65           66 
#> -0.836831742 -0.149611226  0.136601052  0.845623306 -0.051699480 -0.979507366 
#>           67           68           69           70           71           72 
#> -0.892554581  0.164547829  2.058861627 -0.927200740 -1.040977484  1.141182130 
#>           73           74           75           76           77           78 
#>  2.002344293 -1.001455274 -0.274604299  0.136601052 -0.904198589 -0.006306397 
#>           79           80           81           82           83           84 
#>  2.058861626  0.780047359 -0.309429601  3.881268138 -0.136306874  0.714709596 
#>           85           86           87           88           89           90 
#>  1.633671405  1.166234921 -1.072784474 -0.274604300  0.136601052 -0.029287641 
#>           91           92           93           94           95           96 
#> -0.029287641 -0.904198589 -0.892554580 -0.812086474  0.164547829  0.136601053 
#>           97           98           99          100          101          102 
#>  0.845623306 -0.915745726  1.488023281 -0.849002137  0.491768886 -0.836831742 
#>          103          104          105          106          107          108 
#> -0.968428480 -0.812086474  1.141182132 -0.836831742 -0.968428480 -0.812086474 
#>          109          110          111          112          113          114 
#> -0.812086474 -0.812086474  1.141182132 -0.836831742 -0.968428480  0.164547828 
#>          115          116          117          118          119          120 
#> -0.824529234 -0.892554580 -0.812086474  1.141182132  0.109560270 -0.979507365 
#>          121          122          123          124          125          126 
#> -0.892554581  1.141182131 -0.836831742 -0.149611226 -0.824529234 -0.892554580 
#>          127          128          129          130          131          132 
#>  2.117816434  1.015721697 -1.062235555 -0.968428480 -0.812086474 -0.812086474 
#>          133          134          135          136          137          138 
#> -0.812086474  0.164547829 -0.824529234 -0.006306396 -0.824529234 -0.892554580 
#>          139          140          141          142          143          144 
#>  0.164547829 -0.824529234  0.879941788 -0.836831742 -0.149611226  0.136601052