Update and Re-fit a iZIP Model

update (i.e., update.izip) will update and (by-default) re-fit a model. It is identical to update in the stats package.

# S3 method for izip
update(object, formula., ..., evaluate = TRUE)

Arguments

object
an object class 'izip', obtained from a call to glm.izip.

formula.
changes to the existing formula in object -- see update.formula

...
other arguments passed to or from other methods (currently unused).

evaluate
logical; if TRUE evaluate the new call otherwise simply return the call

See also

Examples

# To update the regression formula
data(bioChemist)
#> Warning: data set ‘bioChemist’ not found

## Fit full model
M_bioChem.full <- glm.izip(art ~ ., data = bioChemists)
M_bioChem.full
#> 
#> Call: glm.izip(formula = art ~ ., data = bioChemists)
#> 
#> Linear Model Coefficients:
#> (Intercept)     femWomen   marMarried         kid5          phd         ment  
#>    0.324570    -0.228620     0.158890    -0.189540     0.010118     0.024687  
#> 
#> Baseline Zero-inflation odds (nu): 0.286
#> Degrees of Freedom: 914 Total (i.e. Null);  909 Residual
#> Null Deviance: 1542.813 
#> Residual Deviance: 
#> AIC: 3240.321 
#> 

## Dropping phd
M.bioChem.null <- update(M_bioChem.full, . ~ . - phd)
M.bioChem.null
#> 
#> Call: glm.izip(formula = art ~ fem + mar + kid5 + ment, data = bioChemists)
#> 
#> Linear Model Coefficients:
#> (Intercept)     femWomen   marMarried         kid5         ment  
#>    0.356620    -0.229230     0.156560    -0.189560     0.024839  
#> 
#> Baseline Zero-inflation odds (nu): 0.286
#> Degrees of Freedom: 914 Total (i.e. Null);  910 Residual
#> Null Deviance: 1542.754 
#> Residual Deviance: 
#> AIC: 3238.432 
#>