Extracting the Variance-Covariance Matrix from a iZIP Model Fit

Extracting the Variance-Covariance Matrix from a iZIP Model Fit

# S3 method for izip
vcov(object, ...)

Arguments

object
an object class 'izip' object, obtained from a call to glm.izip

...
other arguments passed to or from other methods (currently unused).

Value

The method will return the estimated covariances between the parameter estimates of the fitted iZIP model.

Examples

data(bioChemists)
M_bioChem <- glm.izip(art~ ., data=bioChemists)
vcov(M_bioChem)

#>               (Intercept)      femWomen    marMarried          kid5
#> (Intercept)  0.0140052010 -2.575516e-03 -3.779114e-03 -1.939299e-04
#> femWomen    -0.0025755158  3.967439e-03  6.131842e-04  4.956370e-04
#> marMarried  -0.0037791143  6.131842e-04  4.984701e-03 -1.279453e-03
#> kid5        -0.0001939299  4.956370e-04 -1.279453e-03  2.143635e-03
#> phd         -0.0029065368  5.885587e-05  2.129996e-04  1.080880e-05
#> ment        -0.0000133543  1.527718e-05  1.268008e-06 -9.800836e-06
#>                       phd          ment
#> (Intercept) -2.906537e-03 -1.335430e-05
#> femWomen     5.885587e-05  1.527718e-05
#> marMarried   2.129996e-04  1.268008e-06
#> kid5         1.080880e-05 -9.800836e-06
#> phd          9.152565e-04 -1.537157e-05
#> ment        -1.537157e-05  4.729492e-06