Extract the Model Frame from an iZIP Model Fit

An accessor function used to extract the model frame from an 'izip' or a 'tsizip' object.

# S3 method for izip
model.frame(formula, ...)

# S3 method for tsizip
model.frame(formula, ...)

Arguments

formula
an object class 'izip' or 'tsizip' object, obtained from a call to glm.izip

...
other arguments passed to or from other methods (currently unused).

Value

The method will return the saved data.frame used when fitting the izip model.

See also

Examples

data(bioChemists)
M_bioChem <- glm.izip(art ~ ., data = bioChemists)
model.frame(M_bioChem)
#>     art   fem     mar kid5   phd ment
#> 1     0   Men Married    0 2.520    7
#> 2     0 Women  Single    0 2.050    6
#> 3     0 Women  Single    0 3.750    6
#> 4     0   Men Married    1 1.180    3
#> 5     0 Women  Single    0 3.750   26
#> 6     0 Women Married    2 3.590    2
#> 7     0 Women  Single    0 3.190    3
#> 8     0   Men Married    2 2.960    4
#> 9     0   Men  Single    0 4.620    6
#> 10    0 Women Married    0 1.250    0
#> 11    0   Men  Single    0 2.960   14
#> 12    0 Women  Single    0 0.755   13
#> 13    0 Women Married    1 3.690    3
#> 14    0 Women Married    0 3.400    4
#> 15    0 Women Married    0 1.790    0
#> 16    0 Women  Single    0 3.090    1
#> 17    0 Women Married    0 2.000    7
#> 18    0   Men Married    2 4.290   13
#> 19    0 Women  Single    0 3.360    7
#> 20    0 Women  Single    0 4.290    9
#> 21    0 Women Married    0 2.260    6
#> 22    0   Men Married    3 2.960    3
#> 23    0   Men Married    1 4.290    5
#> 24    0   Men Married    1 2.860    4
#> 25    0   Men Married    3 2.760    1
#> 26    0 Women Married    0 1.520    3
#> 27    0 Women Married    1 3.540    8
#> 28    0   Men Married    1 4.290    3
#> 29    0 Women  Single    0 3.090    0
#> 30    0   Men Married    0 2.320    3
#> 31    0 Women  Single    0 2.390    0
#> 32    0   Men Married    0 4.290    8
#> 33    0 Women  Single    0 1.505   13
#> 34    0   Men  Single    0 2.100    0
#> 35    0   Men Married    1 1.520    0
#> 36    0 Women  Single    0 2.000    0
#> 37    0 Women  Single    0 1.750    0
#> 38    0   Men  Single    0 4.290    9
#> 39    0   Men Married    1 1.220    2
#> 40    0 Women Married    0 3.750    5
#> 41    0   Men  Single    0 4.290    9
#> 42    0 Women Married    1 2.860    3
#> 43    0   Men  Single    0 3.920    5
#> 44    0 Women Married    1 3.360    3
#> 45    0   Men  Single    0 2.100    6
#> 46    0   Men Married    1 1.180    0
#> 47    0   Men Married    0 2.500    1
#> 48    0   Men Married    0 3.360   25
#> 49    0 Women  Single    0 4.540   14
#> 50    0 Women Married    0 3.920    8
#> 51    0   Men  Single    0 4.540    5
#> 52    0 Women Married    2 1.970    0
#> 53    0   Men Married    2 2.960   12
#> 54    0   Men Married    0 4.290    8
#> 55    0   Men Married    1 2.260    2
#> 56    0 Women  Single    0 2.120   11
#> 57    0 Women  Single    0 3.210    2
#> 58    0   Men Married    1 3.150    1
#> 59    0   Men Married    1 3.920    1
#> 60    0   Men  Single    0 2.960    0
#> 61    0 Women  Single    0 2.860    6
#> 62    0   Men Married    1 2.100    0
#> 63    0 Women  Single    0 3.750    4
#> 64    0 Women Married    0 3.540    2
#> 65    0 Women Married    0 2.540    1
#> 66    0   Men Married    1 2.760    3
#> 67    0   Men Married    0 4.540   10
#> 68    0   Men Married    0 1.680    6
#> 69    0   Men Married    1 2.200    0
#> 70    0 Women  Single    0 1.005    0
#> 71    0   Men  Single    0 2.120    2
#> 72    0 Women Married    2 2.580    0
#> 73    0 Women Married    0 1.790    1
#> 74    0   Men  Single    0 4.290    2
#> 75    0 Women Married    0 1.280    4
#> 76    0   Men Married    2 2.580    5
#> 77    0   Men Married    2 2.120    0
#> 78    0   Men Married    1 2.210    1
#> 79    0   Men Married    1 2.260    5
#> 80    0 Women  Single    0 3.210    5
#> 81    0 Women  Single    0 2.100    2
#> 82    0 Women Married    1 3.540    1
#> 83    0   Men Married    1 3.400   18
#> 84    0   Men Married    0 2.210    6
#> 85    0   Men Married    0 2.210   19
#> 86    0   Men Married    2 0.920    1
#> 87    0 Women  Single    0 4.290   35
#> 88    0   Men Married    2 3.150    6
#> 89    0   Men Married    0 4.290   19
#> 90    0   Men Married    1 2.510    8
#> 91    0   Men Married    1 4.290    1
#> 92    0 Women Married    0 2.760    0
#> 93    0   Men Married    0 1.420    3
#> 94    0   Men Married    0 2.210   19
#> 95    0   Men  Single    0 4.140    5
#> 96    0   Men Married    0 1.520    7
#> 97    0   Men  Single    0 4.620    8
#> 98    0 Women  Single    0 3.620    7
#> 99    0 Women Married    2 3.750    4
#> 100   0   Men Married    0 4.290    8
#> 101   0 Women  Single    0 3.750    5
#> 102   0   Men Married    1 2.100    0
#> 103   0   Men Married    2 1.400    7
#> 104   0 Women Married    0 3.190    4
#> 105   0   Men Married    2 1.810    1
#> 106   0   Men Married    1 3.150    2
#> 107   0 Women Married    0 3.750    7
#> 108   0 Women  Single    0 2.050    6
#> 109   0   Men Married    3 2.210    3
#> 110   0 Women Married    2 3.360    2
#> 111   0 Women Married    0 3.470   11
#> 112   0 Women  Single    0 2.120    0
#> 113   0 Women Married    1 3.360    2
#> 114   0   Men Married    0 2.540    5
#> 115   0   Men Married    2 2.150    4
#> 116   0   Men Married    1 4.540    3
#> 117   0   Men Married    0 1.680    4
#> 118   0 Women Married    1 2.550    0
#> 119   0 Women  Single    0 2.120   11
#> 120   0   Men Married    1 3.470    2
#> 121   0 Women Married    0 3.470    1
#> 122   0 Women  Single    0 3.360    9
#> 123   0   Men Married    1 2.120   10
#> 124   0 Women  Single    0 2.100    1
#> 125   0 Women  Single    0 4.290    1
#> 126   0   Men Married    1 2.860   13
#> 127   0 Women Married    1 4.290    9
#> 128   0   Men Married    0 3.210    2
#> 129   0   Men  Single    0 2.100    0
#> 130   0 Women Married    0 1.810    5
#> 131   0 Women  Single    0 2.520    2
#> 132   0   Men Married    1 3.360   23
#> 133   0 Women Married    0 4.540    4
#> 134   0 Women Married    1 4.290    4
#> 135   0 Women  Single    0 2.120    9
#> 136   0   Men Married    2 4.620   30
#> 137   0 Women Married    1 3.590   14
#> 138   0 Women  Single    0 3.590    4
#> 139   0   Men Married    3 2.860    1
#> 140   0 Women  Single    0 1.505    3
#> 141   0 Women Married    1 2.500    0
#> 142   0 Women Married    0 4.620    0
#> 143   0   Men  Single    0 2.500    0
#> 144   0 Women Married    1 3.360    3
#> 145   0 Women  Single    0 2.000    2
#> 146   0 Women Married    2 3.410    5
#> 147   0   Men Married    0 2.100    0
#> 148   0 Women  Single    0 2.000    2
#> 149   0   Men Married    0 3.360    2
#> 150   0   Men Married    3 3.590    7
#> 151   0   Men  Single    0 2.140    2
#> 152   0 Women Married    0 3.590   12
#> 153   0 Women  Single    0 2.390   10
#> 154   0   Men Married    3 4.540   11
#> 155   0   Men  Single    0 2.960    2
#> 156   0   Men Married    1 2.210    5
#> 157   0   Men Married    2 3.690    5
#> 158   0 Women Married    0 1.780    1
#> 159   0 Women  Single    0 2.860   11
#> 160   0   Men Married    0 2.860    8
#> 161   0 Women Married    0 2.580    3
#> 162   0 Women  Single    0 3.750    1
#> 163   0 Women Married    1 1.220   11
#> 164   0 Women  Single    0 2.100    9
#> 165   0 Women  Single    0 3.590   18
#> 166   0 Women  Single    0 3.540    5
#> 167   0 Women  Single    0 2.860   16
#> 168   0 Women Married    0 1.860    6
#> 169   0   Men Married    0 2.960    0
#> 170   0 Women Married    0 3.690    7
#> 171   0 Women Married    0 3.690    2
#> 172   0 Women Married    0 4.290    1
#> 173   0 Women  Single    0 2.520    1
#> 174   0 Women  Single    0 1.760   10
#> 175   0   Men Married    2 2.320    2
#> 176   0 Women  Single    0 3.690    2
#> 177   0 Women  Single    0 3.750   14
#> 178   0 Women  Single    0 4.290    5
#> 179   0 Women Married    2 4.540    6
#> 180   0   Men Married    1 2.100    0
#> 181   0 Women Married    1 2.540    0
#> 182   0   Men Married    1 2.210   19
#> 183   0   Men  Single    0 4.540    8
#> 184   0   Men Married    1 2.100   12
#> 185   0   Men  Single    0 2.260    5
#> 186   0 Women Married    1 4.620   15
#> 187   0 Women  Single    0 2.760    4
#> 188   0 Women  Single    0 4.290   12
#> 189   0 Women  Single    0 0.755    0
#> 190   0 Women Married    0 1.830    2
#> 191   0 Women  Single    0 1.790    0
#> 192   0 Women Married    1 4.290    6
#> 193   0 Women  Single    0 2.500    0
#> 194   0   Men Married    0 2.140    1
#> 195   0   Men Married    2 3.400   13
#> 196   0 Women  Single    0 2.870   15
#> 197   0 Women Married    0 2.960    0
#> 198   0 Women Married    2 3.190    0
#> 199   0   Men Married    1 3.690   16
#> 200   0 Women Married    2 1.255    0
#> 201   0   Men Married    2 4.140   12
#> 202   0   Men Married    1 2.390    5
#> 203   0   Men Married    2 3.920    2
#> 204   0 Women  Single    0 2.960    4
#> 205   0 Women Married    0 2.120    2
#> 206   0 Women Married    2 2.120    1
#> 207   0 Women Married    0 2.610    3
#> 208   0   Men  Single    0 4.290    0
#> 209   0   Men Married    3 2.860    7
#> 210   0 Women Married    0 2.390    0
#> 211   0   Men Married    2 3.340    1
#> 212   0 Women  Single    0 3.620    8
#> 213   0 Women Married    1 3.590    8
#> 214   0   Men Married    2 4.540    1
#> 215   0 Women Married    0 3.590    1
#> 216   0   Men  Single    0 4.290    9
#> 217   0   Men Married    0 1.520    4
#> 218   0   Men Married    0 4.290    6
#> 219   0 Women Married    0 4.620    5
#> 220   0   Men Married    1 4.290   13
#> 221   0   Men Married    0 2.510    3
#> 222   0   Men  Single    0 2.860    3
#> 223   0 Women Married    0 2.960    8
#> 224   0 Women Married    0 3.690    3
#> 225   0 Women Married    0 3.470    2
#> 226   0   Men Married    2 4.290   15
#> 227   0   Men  Single    0 4.290    0
#> 228   0 Women  Single    0 3.920    1
#> 229   0 Women Married    0 2.000    6
#> 230   0   Men Married    0 2.320    4
#> 231   0 Women Married    1 2.000    9
#> 232   0   Men Married    1 1.680   27
#> 233   0   Men  Single    0 3.470    0
#> 234   0   Men Married    0 2.200    2
#> 235   0   Men Married    1 2.120   10
#> 236   0 Women  Single    0 1.400   14
#> 237   0 Women  Single    0 3.360    3
#> 238   0   Men Married    2 4.290   24
#> 239   0   Men Married    1 2.100    0
#> 240   0   Men  Single    0 4.290    5
#> 241   0 Women  Single    0 4.290    0
#> 242   0 Women  Single    0 2.830    1
#> 243   0   Men Married    1 2.580    6
#> 244   0 Women Married    0 3.150    7
#> 245   0   Men  Single    0 4.290    2
#> 246   0   Men  Single    0 2.960    0
#> 247   0   Men Married    2 3.590    5
#> 248   0 Women  Single    0 2.390   15
#> 249   0 Women  Single    0 1.250    2
#> 250   0 Women  Single    0 2.760    1
#> 251   0   Men Married    0 3.400   13
#> 252   0   Men  Single    0 4.290    7
#> 253   0 Women  Single    0 3.690    4
#> 254   0 Women Married    0 2.860   13
#> 255   0 Women  Single    0 4.290    3
#> 256   0 Women  Single    0 4.140    2
#> 257   0 Women Married    0 2.390    2
#> 258   0   Men Married    2 2.860   11
#> 259   0 Women Married    0 2.960    7
#> 260   0   Men  Single    0 3.590    8
#> 261   0   Men Married    1 1.810    5
#> 262   0   Men Married    1 4.290    2
#> 263   0   Men Married    1 2.100    0
#> 264   0 Women  Single    0 3.190    0
#> 265   0   Men Married    2 4.290    2
#> 266   0   Men Married    0 4.290    2
#> 267   0   Men Married    2 3.400    0
#> 268   0   Men  Single    0 4.290   35
#> 269   0   Men Married    2 2.320    5
#> 270   0   Men  Single    0 1.680    5
#> 271   0 Women Married    0 2.580    0
#> 272   0   Men Married    0 3.690    3
#> 273   0   Men Married    2 3.410    7
#> 274   0   Men Married    0 2.320    0
#> 275   0   Men Married    1 3.690    8
#> 276   1   Men Married    0 2.540    4
#> 277   1 Women Married    1 1.760    4
#> 278   1   Men Married    1 4.290   13
#> 279   1   Men Married    3 2.100    0
#> 280   1 Women Married    1 3.090   12
#> 281   1 Women  Single    0 3.360    6
#> 282   1   Men Married    0 2.100    0
#> 283   1 Women Married    1 3.750    7
#> 284   1 Women  Single    0 2.140    6
#> 285   1 Women  Single    0 4.290   22
#> 286   1 Women  Single    0 2.000    2
#> 287   1   Men Married    0 2.000    4
#> 288   1   Men Married    1 2.000   16
#> 289   1   Men Married    1 2.610    6
#> 290   1   Men Married    0 4.250   12
#> 291   1   Men  Single    0 2.360   14
#> 292   1 Women Married    0 4.290   11
#> 293   1   Men Married    0 2.100    2
#> 294   1 Women Married    0 2.870    5
#> 295   1 Women Married    0 3.470    3
#> 296   1   Men Married    0 1.780    1
#> 297   1   Men Married    0 4.140   15
#> 298   1   Men Married    3 4.290    0
#> 299   1   Men  Single    0 4.290    3
#> 300   1   Men  Single    0 1.810   10
#> 301   1 Women  Single    0 2.320   15
#> 302   1 Women Married    0 2.610    7
#> 303   1   Men Married    0 1.680   34
#> 304   1 Women Married    0 3.190    3
#> 305   1   Men Married    1 3.590    5
#> 306   1   Men Married    0 2.500    0
#> 307   1   Men  Single    0 3.400    6
#> 308   1 Women  Single    0 2.260    4
#> 309   1   Men  Single    0 3.400   12
#> 310   1   Men Married    1 3.920    0
#> 311   1   Men Married    2 2.960    0
#> 312   1   Men  Single    0 4.290   15
#> 313   1   Men Married    0 4.290   11
#> 314   1 Women Married    1 3.320   24
#> 315   1   Men  Single    0 1.860   15
#> 316   1 Women  Single    0 2.870    6
#> 317   1 Women Married    0 1.640   22
#> 318   1 Women  Single    0 2.250    7
#> 319   1 Women  Single    0 2.100    8
#> 320   1 Women Married    0 4.290    6
#> 321   1   Men Married    0 1.750    3
#> 322   1   Men Married    0 2.960    8
#> 323   1 Women  Single    0 2.250    4
#> 324   1   Men Married    2 4.250   20
#> 325   1   Men Married    2 2.560   22
#> 326   1 Women  Single    0 3.190   11
#> 327   1   Men Married    2 3.920    2
#> 328   1   Men Married    1 1.780   77
#> 329   1 Women Married    0 3.400   14
#> 330   1   Men Married    1 1.740    4
#> 331   1   Men  Single    0 4.290    6
#> 332   1 Women  Single    0 2.000    6
#> 333   1 Women Married    0 2.870   18
#> 334   1   Men Married    0 4.290    6
#> 335   1 Women Married    0 3.190    2
#> 336   1   Men Married    1 1.220    1
#> 337   1 Women Married    2 3.540   17
#> 338   1   Men Married    1 4.290    1
#> 339   1   Men Married    1 4.620    3
#> 340   1 Women  Single    0 2.000    6
#> 341   1 Women  Single    0 3.590    3
#> 342   1 Women  Single    0 3.190    6
#> 343   1   Men Married    1 3.400    2
#> 344   1 Women Married    0 1.280    2
#> 345   1   Men Married    2 1.740    4
#> 346   1   Men Married    1 2.260    1
#> 347   1 Women Married    0 3.590    5
#> 348   1   Men Married    1 4.620   16
#> 349   1 Women  Single    0 1.400    2
#> 350   1 Women  Single    0 0.920    4
#> 351   1   Men Married    1 2.390    3
#> 352   1   Men Married    0 2.260    1
#> 353   1   Men Married    2 2.860    7
#> 354   1   Men Married    2 2.760    8
#> 355   1   Men Married    1 1.630    8
#> 356   1   Men Married    2 4.290   12
#> 357   1 Women  Single    0 1.400    3
#> 358   1 Women Married    0 2.120    0
#> 359   1 Women Married    1 2.610    6
#> 360   1   Men Married    0 1.800    1
#> 361   1 Women Married    0 4.290   21
#> 362   1 Women  Single    0 3.690   18
#> 363   1 Women Married    0 4.620   16
#> 364   1   Men  Single    0 2.150   10
#> 365   1   Men Married    2 2.100    0
#> 366   1 Women Married    0 2.000    5
#> 367   1 Women  Single    0 1.255    0
#> 368   1   Men Married    2 1.750    2
#> 369   1 Women  Single    0 3.850    2
#> 370   1   Men  Single    0 4.250    4
#> 371   1 Women  Single    0 3.470    3
#> 372   1   Men Married    1 3.920    8
#> 373   1   Men Married    2 2.510    3
#> 374   1 Women  Single    0 1.400   12
#> 375   1   Men Married    1 2.000   11
#> 376   1 Women  Single    0 2.000   10
#> 377   1   Men  Single    0 4.290    2
#> 378   1   Men  Single    0 3.360   10
#> 379   1   Men Married    0 4.620    1
#> 380   1   Men Married    1 2.560   16
#> 381   1   Men Married    0 4.620   18
#> 382   1   Men Married    0 3.590   10
#> 383   1   Men Married    2 4.290    6
#> 384   1 Women  Single    0 3.590    8
#> 385   1 Women  Single    0 4.290    6
#> 386   1   Men Married    0 3.190    3
#> 387   1 Women Married    0 3.590    1
#> 388   1 Women Married    0 2.960   19
#> 389   1 Women  Single    0 2.580    2
#> 390   1   Men Married    3 2.960    6
#> 391   1 Women Married    0 4.540   13
#> 392   1   Men Married    1 4.540   10
#> 393   1   Men Married    2 2.500    4
#> 394   1   Men Married    0 1.680    8
#> 395   1   Men Married    0 2.320    4
#> 396   1 Women Married    3 3.190    9
#> 397   1 Women  Single    0 2.210    2
#> 398   1 Women Married    0 1.250    5
#> 399   1   Men  Single    0 4.140    5
#> 400   1 Women Married    0 4.620    4
#> 401   1 Women  Single    0 4.540   37
#> 402   1 Women Married    0 3.150    9
#> 403   1   Men  Single    0 4.290   11
#> 404   1 Women Married    0 1.950    3
#> 405   1   Men  Single    0 2.000    9
#> 406   1 Women Married    0 4.540    2
#> 407   1 Women Married    2 2.580    2
#> 408   1 Women Married    0 3.850   29
#> 409   1   Men Married    1 3.410   10
#> 410   1 Women Married    1 2.260    9
#> 411   1 Women  Single    0 4.250    2
#> 412   1   Men Married    1 3.590    6
#> 413   1 Women Married    1 2.120    9
#> 414   1   Men Married    0 2.960    7
#> 415   1 Women  Single    0 2.520    0
#> 416   1 Women  Single    0 2.260    4
#> 417   1 Women  Single    0 4.540    9
#> 418   1 Women Married    2 2.860    7
#> 419   1   Men Married    0 4.540    2
#> 420   1   Men Married    0 3.590   12
#> 421   1 Women Married    0 4.620   18
#> 422   1   Men Married    3 1.670    3
#> 423   1   Men Married    2 1.750    1
#> 424   1 Women Married    1 2.830    5
#> 425   1   Men Married    3 3.150   11
#> 426   1   Men Married    1 3.590    9
#> 427   1 Women  Single    0 3.750   24
#> 428   1 Women  Single    0 2.050   17
#> 429   1   Men Married    0 4.290    0
#> 430   1   Men Married    2 4.290    0
#> 431   1 Women Married    0 2.390    7
#> 432   1   Men Married    1 4.540    3
#> 433   1   Men Married    0 2.320    4
#> 434   1 Women  Single    0 4.290   25
#> 435   1   Men Married    1 3.690    5
#> 436   1   Men  Single    0 4.290    9
#> 437   1   Men  Single    0 3.920    2
#> 438   1   Men Married    0 2.760    5
#> 439   1   Men  Single    0 2.760    6
#> 440   1   Men Married    2 3.620    4
#> 441   1   Men Married    1 4.290   14
#> 442   1   Men  Single    0 2.870   12
#> 443   1 Women Married    1 2.390    9
#> 444   1   Men  Single    0 4.290   14
#> 445   1   Men Married    0 1.780    4
#> 446   1   Men Married    1 2.100    8
#> 447   1 Women Married    0 2.250    3
#> 448   1   Men Married    1 2.260    5
#> 449   1   Men Married    0 1.860    5
#> 450   1 Women  Single    0 3.920    4
#> 451   1   Men Married    0 4.340   10
#> 452   1 Women  Single    0 3.360    4
#> 453   1 Women  Single    0 2.510    1
#> 454   1 Women  Single    0 3.750   21
#> 455   1   Men Married    1 4.540   45
#> 456   1 Women  Single    0 1.630    8
#> 457   1 Women  Single    0 4.620   10
#> 458   1   Men Married    0 1.520    2
#> 459   1   Men Married    0 2.550   11
#> 460   1 Women Married    0 1.180    1
#> 461   1 Women Married    2 3.090    9
#> 462   1 Women  Single    0 4.620    5
#> 463   1 Women Married    0 2.320    1
#> 464   1   Men  Single    0 2.050    3
#> 465   1   Men Married    0 4.290    1
#> 466   1   Men  Single    0 3.400    2
#> 467   1   Men  Single    0 4.290   10
#> 468   1   Men Married    1 3.590    2
#> 469   1   Men  Single    0 3.360    5
#> 470   1 Women  Single    0 2.050   13
#> 471   1 Women Married    0 3.590   32
#> 472   1 Women  Single    0 3.850   18
#> 473   1   Men  Single    0 3.360    0
#> 474   1 Women Married    0 1.750    0
#> 475   1 Women Married    0 3.470    3
#> 476   1   Men Married    1 2.610    4
#> 477   1   Men  Single    0 2.760    4
#> 478   1   Men Married    2 4.290    9
#> 479   1 Women  Single    0 1.400    3
#> 480   1 Women Married    0 4.620   18
#> 481   1   Men Married    1 2.860   17
#> 482   1   Men Married    1 2.860   17
#> 483   1   Men Married    0 2.860   39
#> 484   1   Men  Single    0 2.000    7
#> 485   1 Women Married    0 1.760    8
#> 486   1   Men Married    0 2.500    0
#> 487   1 Women  Single    0 4.290   17
#> 488   1   Men  Single    0 2.560   16
#> 489   1 Women Married    1 3.540   14
#> 490   1 Women Married    0 3.690    1
#> 491   1   Men  Single    0 2.870    5
#> 492   1 Women  Single    0 1.655    0
#> 493   1   Men Married    1 4.340    4
#> 494   1   Men Married    0 3.540   11
#> 495   1   Men Married    2 4.290   10
#> 496   1   Men Married    0 1.810    0
#> 497   1 Women  Single    0 3.400    7
#> 498   1   Men Married    1 1.810    5
#> 499   1 Women  Single    0 3.190   18
#> 500   1 Women Married    1 2.320    5
#> 501   1 Women  Single    0 4.540   19
#> 502   1   Men Married    1 4.290    2
#> 503   1   Men Married    1 2.860    2
#> 504   1 Women Married    1 2.860    2
#> 505   1   Men Married    1 1.220    0
#> 506   1   Men Married    0 2.100    0
#> 507   1   Men Married    1 1.680   12
#> 508   1 Women  Single    0 3.360   34
#> 509   1 Women Married    0 2.870   15
#> 510   1   Men Married    1 3.470    1
#> 511   1   Men  Single    0 2.760    7
#> 512   1   Men Married    1 3.540   21
#> 513   1 Women Married    0 1.950   11
#> 514   1   Men Married    1 3.590   19
#> 515   1 Women  Single    0 3.190    3
#> 516   1 Women  Single    0 2.760    2
#> 517   1 Women Married    2 4.290   13
#> 518   1   Men Married    2 3.190    5
#> 519   1   Men Married    0 2.960    0
#> 520   1   Men Married    1 2.760    7
#> 521   1 Women Married    0 1.810   11
#> 522   2   Men Married    0 3.360    4
#> 523   2   Men  Single    0 3.090   14
#> 524   2 Women  Single    0 3.750    8
#> 525   2 Women Married    0 4.540   10
#> 526   2 Women  Single    0 1.630    6
#> 527   2 Women Married    0 3.090    2
#> 528   2 Women  Single    0 3.690    4
#> 529   2 Women  Single    0 2.120    2
#> 530   2 Women  Single    0 2.260    3
#> 531   2   Men Married    1 4.290    2
#> 532   2 Women Married    1 1.740    4
#> 533   2   Men Married    2 2.960    8
#> 534   2 Women  Single    0 2.580    5
#> 535   2   Men Married    1 4.540   10
#> 536   2 Women  Single    0 4.620   12
#> 537   2   Men Married    1 4.290   13
#> 538   2 Women  Single    0 4.540   13
#> 539   2 Women Married    0 3.590    0
#> 540   2 Women  Single    0 4.290    5
#> 541   2   Men  Single    0 3.690   26
#> 542   2 Women  Single    0 3.210    0
#> 543   2 Women Married    1 3.470   11
#> 544   2   Men  Single    0 4.540   37
#> 545   2   Men Married    0 1.950    4
#> 546   2   Men Married    2 2.960   46
#> 547   2   Men Married    1 4.540   53
#> 548   2   Men Married    0 3.400    9
#> 549   2   Men Married    0 2.100    4
#> 550   2   Men Married    2 4.290    8
#> 551   2 Women  Single    0 3.750    5
#> 552   2 Women  Single    0 2.050    2
#> 553   2   Men Married    1 2.500    2
#> 554   2 Women Married    1 4.620   17
#> 555   2 Women Married    0 1.800   10
#> 556   2 Women  Single    0 3.090    0
#> 557   2   Men  Single    0 4.290   10
#> 558   2   Men Married    2 1.800    6
#> 559   2 Women Married    0 3.590    5
#> 560   2   Men Married    0 2.760    3
#> 561   2   Men Married    0 4.140   16
#> 562   2   Men  Single    0 3.590    3
#> 563   2 Women  Single    0 4.540   12
#> 564   2 Women Married    0 0.920    0
#> 565   2 Women  Single    0 4.290   15
#> 566   2   Men Married    1 2.390    4
#> 567   2   Men Married    0 1.180    0
#> 568   2   Men Married    2 4.540    2
#> 569   2 Women Married    2 3.850   47
#> 570   2   Men Married    2 2.960   21
#> 571   2 Women Married    0 3.360    8
#> 572   2 Women Married    0 3.190    5
#> 573   2 Women Married    1 3.690   15
#> 574   2   Men Married    0 1.630    8
#> 575   2   Men Married    0 1.720    1
#> 576   2 Women Married    0 2.580    5
#> 577   2   Men Married    0 1.520    5
#> 578   2   Men  Single    0 2.860    4
#> 579   2 Women Married    2 1.220    4
#> 580   2 Women  Single    0 2.540    4
#> 581   2 Women  Single    0 3.360    4
#> 582   2   Men Married    2 2.870   20
#> 583   2   Men Married    1 3.150    7
#> 584   2 Women  Single    0 4.620   14
#> 585   2   Men Married    1 1.760    7
#> 586   2   Men Married    1 1.520    3
#> 587   2   Men  Single    0 4.290    5
#> 588   2   Men Married    2 4.290   10
#> 589   2   Men Married    1 2.560    5
#> 590   2   Men  Single    0 4.540   29
#> 591   2 Women Married    0 1.860    5
#> 592   2 Women  Single    0 4.290    1
#> 593   2 Women Married    0 3.920    7
#> 594   2   Men  Single    0 2.260    2
#> 595   2   Men Married    1 4.290   14
#> 596   2 Women Married    1 3.690   10
#> 597   2   Men Married    2 1.970    0
#> 598   2   Men Married    1 3.190   10
#> 599   2 Women  Single    0 4.540    5
#> 600   2   Men  Single    0 4.290   25
#> 601   2   Men Married    0 2.830    1
#> 602   2   Men Married    3 4.340    7
#> 603   2 Women  Single    0 3.360    3
#> 604   2 Women  Single    0 3.410   14
#> 605   2 Women  Single    0 2.860    3
#> 606   2   Men Married    0 4.290   24
#> 607   2   Men Married    1 2.760    8
#> 608   2 Women Married    0 3.750    0
#> 609   2   Men Married    1 4.290   12
#> 610   2   Men Married    2 2.210   12
#> 611   2   Men Married    0 3.540   21
#> 612   2   Men Married    0 2.960    2
#> 613   2 Women Married    0 2.120    2
#> 614   2   Men  Single    0 2.390    4
#> 615   2 Women  Single    0 3.190    5
#> 616   2 Women  Single    0 1.970    8
#> 617   2 Women Married    0 4.540    9
#> 618   2 Women  Single    0 2.550   36
#> 619   2   Men Married    0 1.760    0
#> 620   2   Men Married    2 4.340   17
#> 621   2 Women Married    0 2.520   16
#> 622   2   Men Married    1 2.560    7
#> 623   2   Men  Single    0 4.540   10
#> 624   2 Women Married    0 1.400    0
#> 625   2   Men Married    2 3.320   27
#> 626   2   Men Married    0 2.510    5
#> 627   2 Women Married    2 2.580    3
#> 628   2   Men Married    2 3.540   16
#> 629   2   Men Married    0 4.540   11
#> 630   2   Men  Single    0 4.290   17
#> 631   2   Men Married    0 4.340    5
#> 632   2 Women Married    1 3.470    4
#> 633   2 Women  Single    0 3.470    5
#> 634   2   Men Married    2 2.500    5
#> 635   2   Men Married    3 2.000    8
#> 636   2   Men  Single    0 3.920   10
#> 637   2 Women Married    0 4.620   23
#> 638   2 Women  Single    0 3.090    6
#> 639   2 Women  Single    0 2.320    4
#> 640   2 Women  Single    0 4.290   21
#> 641   2 Women  Single    0 3.360    7
#> 642   2   Men Married    0 3.690    5
#> 643   2   Men Married    0 4.290   17
#> 644   2 Women  Single    0 3.590   11
#> 645   2   Men Married    0 4.290   13
#> 646   2 Women Married    1 3.210    7
#> 647   2 Women  Single    0 2.260    8
#> 648   2   Men Married    0 2.500    0
#> 649   2   Men  Single    0 2.260    3
#> 650   2 Women Married    0 3.750    4
#> 651   2 Women  Single    0 4.290    4
#> 652   2   Men Married    0 3.190   11
#> 653   2 Women Married    1 2.000    7
#> 654   2 Women Married    0 3.090    1
#> 655   2 Women Married    1 3.590    2
#> 656   2 Women Married    0 4.340   11
#> 657   2 Women  Single    0 2.870    9
#> 658   2   Men Married    0 2.100    2
#> 659   2 Women  Single    0 2.830   26
#> 660   2   Men Married    1 2.100    3
#> 661   2   Men Married    1 1.780    1
#> 662   2   Men Married    0 2.510    0
#> 663   2   Men Married    1 2.520   10
#> 664   2 Women Married    1 3.400    4
#> 665   2 Women Married    1 3.090    3
#> 666   2   Men Married    0 4.290   13
#> 667   2 Women Married    0 2.960    8
#> 668   2 Women Married    2 3.590   11
#> 669   2   Men  Single    0 4.540   55
#> 670   2 Women Married    0 3.590   11
#> 671   2   Men Married    1 3.420    3
#> 672   2 Women Married    1 4.290    4
#> 673   2 Women  Single    0 3.190    4
#> 674   2 Women Married    1 4.290   11
#> 675   2   Men Married    0 4.620    3
#> 676   2   Men Married    2 1.860    6
#> 677   2   Men Married    0 4.620    8
#> 678   2 Women Married    2 4.540   48
#> 679   2 Women Married    0 3.750    8
#> 680   2   Men Married    0 1.760    0
#> 681   2 Women Married    0 3.410   31
#> 682   2   Men Married    0 2.550   24
#> 683   2   Men Married    1 1.250    2
#> 684   2 Women Married    0 1.480    9
#> 685   2 Women Married    0 2.830    3
#> 686   2 Women Married    0 4.290    6
#> 687   2 Women  Single    0 3.190   10
#> 688   2 Women  Single    0 1.400    0
#> 689   2 Women Married    2 3.210    2
#> 690   2   Men  Single    0 4.620    9
#> 691   2 Women Married    0 1.450    7
#> 692   2 Women Married    1 3.340    4
#> 693   2 Women  Single    0 4.540   23
#> 694   2   Men Married    2 4.290   30
#> 695   2 Women  Single    0 3.690    7
#> 696   2 Women  Single    0 2.580    2
#> 697   2 Women Married    1 4.540    7
#> 698   2   Men Married    1 4.540    6
#> 699   2   Men  Single    0 4.290    9
#> 700   3 Women  Single    0 4.290   12
#> 701   3 Women  Single    0 3.750    0
#> 702   3   Men  Single    0 4.620    8
#> 703   3   Men Married    2 4.540   31
#> 704   3   Men Married    0 3.090    5
#> 705   3   Men Married    1 1.950   14
#> 706   3 Women Married    1 3.590    3
#> 707   3   Men Married    1 2.610    5
#> 708   3   Men  Single    0 2.870   12
#> 709   3   Men  Single    0 4.290   30
#> 710   3 Women Married    0 3.410   19
#> 711   3 Women Married    1 2.320    5
#> 712   3 Women Married    0 2.120    3
#> 713   3   Men Married    0 4.540   37
#> 714   3   Men Married    0 2.960    8
#> 715   3   Men  Single    0 2.870   12
#> 716   3   Men Married    0 2.320    4
#> 717   3   Men Married    2 1.800    6
#> 718   3 Women  Single    0 2.390   11
#> 719   3 Women Married    1 4.620   14
#> 720   3 Women  Single    0 4.290   10
#> 721   3   Men Married    0 4.540    5
#> 722   3   Men  Single    0 3.590   16
#> 723   3   Men Married    2 4.290    0
#> 724   3   Men Married    1 2.960    5
#> 725   3   Men Married    2 4.290   25
#> 726   3   Men Married    1 4.290   18
#> 727   3   Men Married    1 2.540    6
#> 728   3   Men  Single    0 3.920    0
#> 729   3   Men  Single    0 2.320    8
#> 730   3 Women  Single    0 4.620   22
#> 731   3 Women Married    0 3.750    6
#> 732   3   Men Married    2 1.950    3
#> 733   3   Men Married    1 1.800    4
#> 734   3   Men Married    0 2.760    8
#> 735   3   Men  Single    0 3.590    7
#> 736   3   Men Married    0 4.250   20
#> 737   3   Men Married    2 2.830    7
#> 738   3 Women  Single    0 2.520    4
#> 739   3 Women Married    0 1.180    0
#> 740   3 Women Married    0 3.360   11
#> 741   3 Women Married    0 4.290   21
#> 742   3 Women Married    1 4.620   49
#> 743   3   Men  Single    0 2.830    8
#> 744   3   Men Married    0 3.690   24
#> 745   3   Men Married    0 2.260   12
#> 746   3 Women  Single    0 2.050    5
#> 747   3   Men Married    1 4.290   19
#> 748   3   Men Married    0 1.420    3
#> 749   3 Women Married    2 1.890   16
#> 750   3   Men Married    0 3.690   26
#> 751   3 Women Married    0 2.960   22
#> 752   3 Women Married    1 1.860    1
#> 753   3 Women Married    0 2.260    6
#> 754   3   Men Married    0 4.290   15
#> 755   3 Women Married    0 3.150    4
#> 756   3   Men Married    0 2.610    0
#> 757   3 Women  Single    0 4.540    3
#> 758   3   Men Married    1 2.210   10
#> 759   3   Men Married    0 1.780    0
#> 760   3   Men Married    1 3.620    5
#> 761   3   Men Married    0 3.920    0
#> 762   3 Women Married    0 4.620   10
#> 763   3   Men Married    0 2.610    5
#> 764   3 Women Married    1 3.360    3
#> 765   3 Women Married    0 3.690   12
#> 766   3   Men Married    0 3.470   14
#> 767   3   Men Married    0 1.250    4
#> 768   3   Men Married    0 4.620   18
#> 769   3   Men Married    1 3.590   14
#> 770   3 Women  Single    0 3.360    2
#> 771   3   Men Married    0 4.290   38
#> 772   3   Men  Single    0 3.540   12
#> 773   3   Men Married    2 2.580    2
#> 774   3   Men  Single    0 1.520    3
#> 775   3   Men Married    0 3.470    8
#> 776   3 Women Married    0 4.620   14
#> 777   3   Men  Single    0 3.540   25
#> 778   3   Men Married    2 2.100    4
#> 779   3   Men Married    2 1.380    8
#> 780   3   Men Married    1 3.090    5
#> 781   3   Men Married    3 2.860    3
#> 782   3 Women  Single    0 4.290    5
#> 783   3 Women Married    1 1.220    3
#> 784   4   Men Married    0 2.830    5
#> 785   4 Women Married    0 4.540   16
#> 786   4   Men Married    0 4.290   22
#> 787   4   Men  Single    0 4.290    7
#> 788   4   Men Married    0 3.540    9
#> 789   4   Men  Single    0 2.520    6
#> 790   4 Women  Single    0 1.400    3
#> 791   4 Women Married    0 4.340    6
#> 792   4 Women Married    1 2.510    2
#> 793   4   Men Married    0 4.290    2
#> 794   4   Men Married    1 2.960    0
#> 795   4   Men  Single    0 3.850   47
#> 796   4   Men  Single    0 4.540    5
#> 797   4 Women Married    0 3.400    5
#> 798   4   Men  Single    0 3.410   21
#> 799   4   Men Married    1 2.960   57
#> 800   4 Women Married    0 3.540    5
#> 801   4   Men Married    2 1.720    2
#> 802   4   Men Married    2 2.860   15
#> 803   4   Men Married    2 4.540   66
#> 804   4   Men Married    1 1.760   11
#> 805   4   Men Married    2 1.860   15
#> 806   4   Men Married    0 3.470    4
#> 807   4   Men  Single    0 3.410    7
#> 808   4   Men Married    0 1.810   10
#> 809   4   Men Married    2 2.050   13
#> 810   4   Men  Single    0 4.620   18
#> 811   4 Women Married    0 2.260    3
#> 812   4   Men Married    1 1.860   47
#> 813   4 Women Married    1 4.290    4
#> 814   4   Men Married    2 3.540    2
#> 815   4   Men Married    0 4.290   30
#> 816   4   Men Married    0 2.540   11
#> 817   4 Women Married    0 4.250   13
#> 818   4 Women  Single    0 2.500   10
#> 819   4 Women Married    1 3.540   12
#> 820   4 Women Married    0 3.190    8
#> 821   4   Men Married    2 4.540   21
#> 822   4 Women  Single    0 2.860   39
#> 823   4 Women Married    0 2.580   14
#> 824   4 Women Married    0 1.280    4
#> 825   4 Women Married    0 3.210   14
#> 826   4 Women  Single    0 2.870    6
#> 827   4 Women  Single    0 3.340   11
#> 828   4 Women  Single    0 3.150    8
#> 829   4 Women  Single    0 2.320    8
#> 830   4   Men  Single    0 4.290   19
#> 831   4   Men  Single    0 3.690    7
#> 832   4   Men Married    1 2.550   21
#> 833   4 Women  Single    0 4.620    3
#> 834   4 Women Married    0 2.360    5
#> 835   4   Men Married    1 4.250   29
#> 836   4   Men Married    1 4.290    5
#> 837   4 Women Married    0 3.190    7
#> 838   4 Women  Single    0 3.190    3
#> 839   4 Women  Single    0 2.560   23
#> 840   4   Men Married    0 1.760    7
#> 841   4   Men Married    1 2.580   25
#> 842   4 Women Married    0 3.410    4
#> 843   4 Women  Single    0 3.590    9
#> 844   4 Women Married    0 1.780    4
#> 845   4 Women  Single    0 3.750    3
#> 846   4   Men Married    2 3.690   16
#> 847   4   Men Married    0 3.620    5
#> 848   4   Men Married    0 1.505    4
#> 849   4   Men Married    2 2.580    6
#> 850   4   Men  Single    0 2.870   15
#> 851   5 Women Married    0 2.260   16
#> 852   5 Women  Single    0 3.750   13
#> 853   5   Men Married    2 3.850   15
#> 854   5   Men Married    1 3.150    5
#> 855   5   Men Married    1 3.150   27
#> 856   5   Men Married    0 4.340    6
#> 857   5   Men Married    0 4.250   12
#> 858   5   Men  Single    0 3.360   11
#> 859   5 Women  Single    0 3.590   14
#> 860   5   Men Married    1 2.580   25
#> 861   5   Men Married    1 3.590   21
#> 862   5   Men Married    1 2.960    4
#> 863   5   Men Married    0 4.540   13
#> 864   5   Men Married    0 4.290   30
#> 865   5 Women  Single    0 2.870   12
#> 866   5 Women Married    0 4.540   15
#> 867   5 Women Married    1 3.850    1
#> 868   5   Men Married    0 2.500    4
#> 869   5   Men Married    1 3.470    0
#> 870   5 Women Married    0 2.860   24
#> 871   5 Women Married    0 4.620    0
#> 872   5   Men Married    1 4.540   53
#> 873   5   Men Married    0 1.250    2
#> 874   5 Women Married    0 3.210   19
#> 875   5   Men Married    1 1.760   11
#> 876   5 Women  Single    0 2.580    6
#> 877   5 Women  Single    0 3.620    3
#> 878   6   Men Married    1 4.620    8
#> 879   6 Women Married    0 2.100   36
#> 880   6 Women Married    2 1.860   38
#> 881   6   Men Married    0 4.340    9
#> 882   6 Women Married    0 4.290   24
#> 883   6   Men Married    2 2.510   11
#> 884   6   Men Married    1 2.960   13
#> 885   6   Men  Single    0 4.290   18
#> 886   6   Men  Single    0 3.400   14
#> 887   6 Women  Single    0 4.540   12
#> 888   6   Men Married    1 3.850   16
#> 889   6 Women  Single    0 3.150    9
#> 890   6 Women  Single    0 4.540   15
#> 891   6   Men  Single    0 3.470    6
#> 892   6 Women Married    0 4.290    1
#> 893   6   Men  Single    0 1.970    4
#> 894   6 Women  Single    0 3.320    6
#> 895   7   Men Married    0 3.590    1
#> 896   7   Men  Single    0 2.540    6
#> 897   7   Men  Single    0 3.410   20
#> 898   7   Men Married    1 1.970    0
#> 899   7 Women  Single    0 3.150    9
#> 900   7   Men  Single    0 4.620   15
#> 901   7   Men  Single    0 4.540   42
#> 902   7   Men Married    0 3.690    9
#> 903   7   Men  Single    0 4.340   19
#> 904   7   Men  Single    0 4.290   19
#> 905   7   Men Married    1 3.590   27
#> 906   7   Men  Single    0 3.690   19
#> 907   8   Men Married    0 2.510   11
#> 908   9   Men Married    1 2.960   23
#> 909   9   Men Married    1 1.860   47
#> 910  10 Women Married    0 3.590   18
#> 911  11   Men Married    2 2.860    7
#> 912  12   Men Married    1 4.290   35
#> 913  12   Men Married    1 1.860    5
#> 914  16   Men Married    0 1.740   21
#> 915  19   Men Married    0 1.860   42