Random Normal Probability Integral Transform

A function to create the normal conditional (randomized) quantile residuals. The majority of the code and descriptions are taken from Dunsmuir and Scott (2015).

izipnormRandPIT(object)

Arguments

object
an object class "izip", obtained from a call to glm.izip.

Value

A list consisting of two elements:

rt

the normal conditional randomized quantile residuals

rdMid

the midpoints of the predictive probability intervals

Details

The function izipPredProb produces the non-randomized probability integral transform(PIT). It returns estimates of the cumulative predictive probabilities as upper and lower bounds of a collection of intervals. If the model is correct, a histogram drawn using these estimated probabilities should resemble a histogram obtained from a sample from the uniform distribution.

This function aims to produce observations which instead resemble a sample from a normal distribution. Such a sample can then be examined by the usual tools for checking normality, such as histograms and normal Q-Q plots.

For each of the intervals produced by izipnormRandPIT, a random uniform observation is generated, which is then converted to a normal observation by applying the inverse standard normal distribution function (using qnorm). The vector of these values is returned by the function in the list element rt. In addition non-random observations which should appear similar to a sample from a normal distribution are obtained by applying qnorm to the mid-points of the predictive distribution intervals. The vector of these values is returned by the function in the list element rtMid.

References

Berkowitz, J. (2001). Testing density forecasts, with applications to risk management. Journal of Business \& Economic Statistics, 19, 465--474.

Dunn, P. K. and Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236--244.

Dunsmuir, W.T.M. and Scott, D.J. (2015). The glarma Package for Observation-Driven Time Series Regression of Counts. Journal of Statistical Software, 67, 1--36.

Examples

data(bioChemists)
M_bioChem <- glm.izip(art ~ ., data = bioChemists)
izipnormRandPIT(M_bioChem)
#> $rt
#>   [1] -0.973224780 -1.572371435 -2.285868417 -1.321877056 -1.519245039
#>   [6] -1.368465180 -1.068915119 -2.017989888 -1.315260435 -0.461738940
#>  [11] -1.615662575 -0.890585765 -0.623718845 -1.585805834 -0.462423160
#>  [16] -0.597347198 -2.204851347 -1.090653374 -0.768491501 -0.825534054
#>  [21] -2.383471639 -1.315358396 -1.412099970 -0.931903141 -0.811519697
#>  [26] -1.121688981 -0.753741489 -0.627316643 -1.487198285 -1.663746551
#>  [31] -0.643418246 -1.355797783 -0.935405891 -0.842391924 -1.876550085
#>  [36] -0.542624120 -0.536109967 -0.783714060 -0.552498478 -1.236362191
#>  [41] -1.276829059 -1.199605451 -1.703567043 -0.866108097 -1.155528688
#>  [46] -0.684684921 -1.659393114 -1.647536596 -2.131656473 -1.345909080
#>  [51] -0.816238815 -1.146026030 -1.012786057 -1.563718731 -0.930131381
#>  [56] -1.597748786 -0.404278574 -0.981183448 -0.984646315 -1.382274755
#>  [61] -1.050983692 -1.193072330 -2.341169543 -1.071625672 -1.157215835
#>  [66] -2.529040642 -1.552786985 -1.217446102 -0.618027549 -1.040825164
#>  [71] -1.034057645 -0.598060508 -1.381759376 -1.401740235 -1.072432617
#>  [76] -0.500937712 -1.026349182 -1.515173560 -0.911202572 -2.060431693
#>  [81] -0.345658322 -1.583373321 -1.382453700 -0.964645891 -1.434977492
#>  [86] -1.336528764 -1.608918469 -0.610765402 -2.770270744 -1.204564624
#>  [91] -0.708904716 -0.629804529 -1.336073902 -1.995306613 -1.250177907
#>  [96] -1.133471556 -1.202446813 -0.809145824 -0.789677048 -1.050705240
#> [101] -0.717310096 -0.600909701 -2.678844263 -0.654864266 -0.328471816
#> [106] -1.052916087 -1.318779769 -0.653815830 -0.688361139 -0.589395303
#> [111] -0.856246057 -0.781594983 -1.315556475 -0.884983723 -0.525915576
#> [116] -1.481224978 -1.360229740 -0.518616945 -1.820458998 -1.583251402
#> [121] -1.629516077 -0.844493750 -0.858735107 -0.393270114 -0.832397787
#> [126] -1.195052881 -0.800825316 -1.593464123 -2.211519153 -0.646127164
#> [131] -1.947217929 -1.148364263 -0.939998208 -1.549396853 -0.591275855
#> [136] -2.593395517 -1.000367178 -1.396330401 -0.268328087 -0.541134920
#> [141] -0.598518150 -1.475821742 -0.774781698 -1.412472171 -0.965749647
#> [146] -1.092703035 -1.077607737 -1.326330927 -1.253279456 -0.386626078
#> [151] -0.590663799 -0.916024096 -0.623492854 -0.844485294 -0.778434235
#> [156] -0.617541039 -0.721319035 -1.485589365 -1.384529270 -1.126570712
#> [161] -0.645147933 -1.353703571 -2.998824579 -0.560350774 -1.294423043
#> [166] -0.670983420 -0.987245084 -1.212847759 -1.366738336 -0.868156008
#> [171] -0.834524223 -1.935352618 -0.469368201 -1.997372037 -0.499181807
#> [176] -1.770697254 -1.154644923 -1.391451572 -0.599163051 -1.632618493
#> [181] -0.321446770 -1.639432599 -1.393666020 -0.885790173 -0.791089299
#> [186] -1.948598524 -0.485788154 -0.807982094 -1.041759948 -0.494162499
#> [191] -1.106493776 -0.477837445 -1.388354057 -2.424456653 -1.772035649
#> [196] -0.792228054 -1.482674652 -1.305767968 -2.461617870 -0.159390707
#> [201] -1.507563716 -0.805664376 -1.266168491 -0.517767720 -1.331059036
#> [206] -0.142200107 -0.928781972 -0.759452369 -0.465985600 -0.678334741
#> [211] -1.022746370 -1.721908387 -1.946585058 -0.712574286 -0.883418939
#> [216] -1.064098689 -1.466977886 -0.888479655 -0.853922000 -0.948666909
#> [221] -0.781200254 -1.555782584 -0.956150769 -1.199962624 -2.196085770
#> [226] -1.037261465 -0.929081528 -1.037478511 -0.717730572 -1.223716411
#> [231] -0.497395605 -2.914523564 -0.839886605 -1.368902739 -0.809037935
#> [236] -0.918690788 -0.723886958 -1.130954908 -0.953734712 -1.552385636
#> [241] -0.873955449 -1.014338433 -0.881757780 -0.963497926 -1.226453633
#> [246] -0.561060258 -1.287424436 -2.236357282 -0.797574997 -1.397649126
#> [251] -2.101075034 -0.770157763 -0.598861718 -0.985886429 -0.892587518
#> [256] -0.855719219 -1.886509349 -1.229370118 -1.690919599 -1.001339417
#> [261] -0.903657722 -0.604522701 -1.542390657 -0.338528762 -1.091860016
#> [266] -1.057267333 -0.507171780 -2.678936335 -0.926930050 -0.844214943
#> [271] -1.072723041 -1.175332973 -1.520541991 -2.425106409 -0.732378957
#> [276] -0.655500905  0.215279516 -0.245078929  0.497804804 -0.123033786
#> [281]  0.209732365 -0.653637939 -0.346761191 -0.365093236 -0.390290442
#> [286] -0.040901993 -0.737047659 -0.811282802  0.016937175 -0.879944302
#> [291] -0.423306174 -0.594418654 -0.335916592  0.012973284 -0.436274550
#> [296] -0.504927923 -0.920248208 -0.143848345 -0.008234113 -0.466585831
#> [301] -0.130151376 -0.111678303 -1.427009720  0.166165531 -0.356722921
#> [306] -0.553499167 -0.128787360 -0.367206156 -0.299879425 -0.376543506
#> [311] -0.290855479 -0.405443021 -0.460240311 -0.239896925 -0.633300209
#> [316] -0.311723111 -0.611726624  0.122595450 -0.457076199 -0.093906653
#> [321] -0.591935883 -0.902774779 -0.292913001 -0.347121992 -0.413454710
#> [326] -0.324986082 -0.080121464 -3.127177042 -0.332373923  0.076330877
#> [331] -0.354656025  0.001367165 -0.804104375 -0.463082665  0.139848200
#> [336] -0.066656950  0.098889095 -0.099261291 -0.028103867 -0.140141197
#> [341]  0.243524632  0.093053170 -0.309474321 -0.401801525 -0.316720697
#> [346]  0.065822809 -0.302621558 -0.901718180  0.396720249  0.066346077
#> [351] -0.448581929 -0.316877736 -0.380264642  0.193342996 -0.648254007
#> [356] -0.492217674  0.331811112  0.258191277  0.107547787 -0.085167007
#> [361] -0.399992434 -0.734164540 -0.572495011 -0.317456758  0.439287184
#> [366] -0.332806326  0.347081656  0.051250098  0.283851622 -0.262346377
#> [371]  0.259344338 -0.109076134 -0.120217672 -0.578730704 -0.610974586
#> [376]  0.143328144 -0.376128517 -0.182744831 -0.071016297 -0.917558133
#> [381] -0.854321666 -0.809316690  0.005678221  0.027672179  0.174879393
#> [386] -0.488747701  0.158655243 -0.391377112  0.189725431 -0.245606930
#> [391] -0.558129706 -0.242853931  0.267446779 -0.738941782 -0.445327242
#> [396]  0.143799662 -0.052299801 -0.395365058 -0.660915924 -0.299537140
#> [401] -0.983342137 -0.091147655 -0.723574252 -0.238225997 -0.547699918
#> [406] -0.153843475 -0.134256793 -1.223909209 -0.074281840  0.199933095
#> [411] -0.268727812 -0.563327360 -0.205915213 -0.538830263  0.279439720
#> [416] -0.028680298 -0.455864409  0.447817125 -0.261086616 -0.676732920
#> [421] -0.572167603 -0.005463381  0.200458896  0.320761529  0.093526246
#> [426] -0.172719692 -0.873934470 -0.097241971 -0.123341044  0.209884975
#> [431] -0.350152331 -0.201599494 -0.446588728 -0.889611356 -0.249647238
#> [436] -0.442696370 -0.183318487 -0.362806378 -0.553414555 -0.191918064
#> [441] -0.272796094 -0.419794869 -0.019835624 -0.716785096 -0.545050975
#> [446] -0.434410011  0.027580298 -0.605322870 -0.765272219 -0.208657812
#> [451] -0.566913426 -0.178654336 -0.251847304 -0.313688979 -1.492062290
#> [456] -0.181615940 -0.514806626 -0.197358093 -0.616146180 -0.254667070
#> [461] -0.112983415 -0.037528969 -0.322382493 -0.123986667 -0.201224771
#> [466] -0.374451338 -0.197527761  0.060148909 -0.621431800 -0.325640360
#> [471] -1.247167872 -0.125422235  0.121129493 -0.297374677  0.086056906
#> [476] -0.581347903  0.002595529 -0.248914104 -0.301157097 -0.499008164
#> [481] -0.334597631 -0.382264229 -1.665628946 -0.259712424 -0.298450539
#> [486] -0.063612441 -0.261329093 -0.660923336 -0.217056016 -0.303824509
#> [491] -0.495372874 -0.241065916 -0.377167302 -0.355830712 -0.475785947
#> [496] -0.167132521 -0.302489785 -0.504832781 -0.113855810  0.270359695
#> [501] -0.640130856 -0.362691181 -0.095580188  0.005422176 -0.240147668
#> [506] -0.602020288 -0.585378350 -0.788495242 -0.214590573 -0.503923258
#> [511] -0.296679876 -0.510380581 -0.666643922 -0.858911640 -0.010145661
#> [516] -0.152908464 -0.103774447 -0.106371745 -0.549460110 -0.609235949
#> [521] -0.711741423  0.331713918  0.379060069  0.542302594  0.063784149
#> [526]  0.330285132  0.488886187  0.709385805  0.759856331  0.836913941
#> [531]  0.473209307  0.679098760  0.225801138  0.946697356  0.193129126
#> [536]  0.731637505  0.161715488  0.282591276  0.565803022  0.337254847
#> [541] -0.065479960  0.837282610  0.817134768 -0.756780449  0.371296481
#> [546] -0.717529140 -1.311773696  0.115095410  0.126552642  0.415725271
#> [551]  0.841849952  0.669652170  0.280462417  0.602153507  0.441422590
#> [556]  0.455422730  0.267886060  0.814246502  0.519251716  0.357517929
#> [561]  0.004297403  0.266758471  0.197444464  0.355924930  0.370397630
#> [566]  0.660937330  0.615958457  0.780966858 -0.532839145 -0.164183520
#> [571]  0.299255413  0.579451997  0.303251547  0.115635469  0.084219379
#> [576]  0.159140195  0.295863183  0.391747619  1.225570343  0.882490211
#> [581]  0.368970450  0.339900911  0.069320142  0.114769001  0.549968361
#> [586]  0.499760930  0.489558147  0.354302910  0.727090700 -0.514647005
#> [591]  0.358436798  0.802645552  0.145587940  0.565193327  0.330606584
#> [596]  0.731720309  1.019498760  0.076534261  0.643490473 -0.077502269
#> [601]  0.185875417  0.916085508  0.816116540  0.590355829  1.016143024
#> [606] -0.754113460  0.621061062  0.660330715 -0.110203479  0.399213897
#> [611] -0.511717468  0.321825786  0.294500731  0.471950731  0.750270322
#> [616]  0.613099667  0.013466531 -0.248927029  0.423899692  0.215154890
#> [621]  0.122661329  0.449851970  0.431437033  0.744901467 -0.118667445
#> [626]  0.122970250  0.677405564  0.254120783 -0.003304068 -0.302122260
#> [631]  0.296698746  0.694988381  0.693073220  0.383581383  1.051943677
#> [636]  0.288347474 -0.127281921  0.411276591  1.000260391 -0.107953284
#> [641]  0.290325304  0.328092976 -0.159063406  0.288638124  0.139024322
#> [646]  0.371526116  0.629162674  0.131524620  0.579533601  0.688895361
#> [651]  0.870841797  0.153589399  0.313206090  0.744465351  0.994488587
#> [656]  0.306043956  0.459255431  0.540497670  0.065816939  0.413965244
#> [661]  0.791150585  0.492318979  0.017716624  0.688731329  0.467845353
#> [666] -0.108339540  0.375502100  0.519518897 -1.466892895  0.548895117
#> [671]  0.381080500  0.663492648  0.988463683  0.373311267  0.173367769
#> [676]  0.412633383 -0.215201506 -0.569664686  0.108315892  0.481798417
#> [681] -0.153656627 -0.387022763  0.471808793  0.360477171  0.384507621
#> [686]  0.472789443  0.440500975  1.040638072  1.114993000 -0.009594263
#> [691]  0.257971136  0.805279588  0.032465479 -0.064142389  0.681964945
#> [696]  0.432665339  0.901895548  0.182376468  0.497999253  1.004320119
#> [701]  1.665808007  0.800649038  0.273090931  0.730847152  1.011195754
#> [706]  1.201479243  0.879984182  0.949665202 -0.111170068  0.764176012
#> [711]  1.194580642  0.974905788 -0.353359312  0.450028035  0.904623507
#> [716]  0.707872640  1.525544076  0.859794684  0.951984420  1.076612007
#> [721]  0.930309260  0.338785157  1.115529076  1.170701176  0.412022288
#> [726]  0.319901252  1.196010635  0.874086429  0.826281141  0.978694620
#> [731]  1.037399980  1.445629532  0.896208962  0.722121195  0.705132828
#> [736]  0.165421165  0.985554335  1.065005722  1.132448296  1.134149189
#> [741]  0.721648146 -0.436777040  0.950876920 -0.157868587  0.420031642
#> [746]  1.309415417  0.483980209  0.955124135  1.439159465 -0.249166962
#> [751]  0.556681180  1.614484538  0.940381296  0.451060432  1.041887967
#> [756]  0.733224653  1.271638568  0.654616382  1.009611816  0.799949495
#> [761]  0.898025946  0.645825046  0.752172256  1.083959306  1.008964162
#> [766]  0.314090491  0.665106604  0.210163051  0.770050762  1.623389792
#> [771] -0.656029917  0.793171315  1.519208349  0.844811734  0.930395917
#> [776]  0.515012483  0.120716298  1.337648248  1.149244156  0.853608892
#> [781]  1.844742487  1.304290308  1.131424443  1.361186090  1.426492625
#> [786]  0.864977091  1.381524527  1.138933346  1.407109384  1.760525515
#> [791]  1.840674858  1.861450692  1.233805848  1.831918653 -0.363775825
#> [796]  1.356968148  1.798433367  0.863368605 -0.855181827  1.383046680
#> [801]  1.724939668  1.258372749 -0.647653740  1.622622763  1.795784386
#> [806]  1.270180421  1.324714592  1.124834600  1.349908645  0.806234374
#> [811]  1.665991710 -0.224744255  1.735035485  1.895936941  0.168802442
#> [816]  1.151456783  1.226549641  1.918374543  1.864937599  1.314356500
#> [821]  1.376607519  0.637401359  1.411511190  1.761611128  1.089431501
#> [826]  1.732728604  1.791106805  1.736395781  1.758771262  1.215564424
#> [831]  1.551628097  1.250695418  1.800956145  1.503242722  0.802418301
#> [836]  1.589986368  1.374235152  1.805443379  0.992961375  1.196913783
#> [841]  0.987643952  1.794684137  1.545846833  1.792328011  1.647017085
#> [846]  1.395671345  1.296151189  1.309454447  1.945170014  1.416362595
#> [851]  1.833765776  1.896186424  1.850299712  2.051956689  1.351091092
#> [856]  1.759979166  1.722345653  1.720104451  1.910322894  1.430365712
#> [861]  1.396345069  2.091702366  1.319196237  0.847369106  2.022891674
#> [866]  1.918750861  2.502878215  2.105557787  2.049708587  1.506515194
#> [871]  2.099510239 -0.048669154  2.067448626  1.531988096  1.718798775
#> [876]  2.138937905  2.322010020  2.456420569  1.475436462  2.060651827
#> [881]  2.053230889  1.726930011  2.480217370  2.402617158  1.873875818
#> [886]  2.131674553  2.821131009  2.254077845  2.754374004  2.399225082
#> [891]  2.741171851  2.594087889  2.631181041  3.096659222  2.767521720
#> [896]  2.825865382  2.234734091  3.232115866  3.149220906  2.768946182
#> [901]  1.162204213  2.737073031  2.408671306  2.293618597  2.079955627
#> [906]  2.484457664  3.109222975  3.234157229  1.925991827  3.719602361
#> [911]  5.066672184  3.786052142  5.063622370  5.263571003  5.123475764
#> 
#> $rtMid
#>   [1] -1.313810491 -0.971792888 -0.984083730 -1.056803196 -1.405256362
#>   [6] -0.776614822 -0.928284122 -0.949480995 -1.169072890 -0.973693821
#>  [11] -1.331224798 -1.091549629 -0.910991654 -1.063012717 -0.977589625
#>  [16] -0.894517775 -1.110173316 -1.127026692 -0.999093503 -1.042784600
#>  [21] -1.092584743 -0.810593364 -1.120562826 -1.089246445 -0.780103831
#>  [26] -1.029777540 -0.996078296 -1.081209248 -0.878381786 -1.222083058
#>  [31] -0.873801028 -1.354525240 -1.097566189 -1.030931090 -1.003871969
#>  [36] -0.871258213 -0.869631702 -1.230167026 -1.038319650 -1.085001945
#>  [41] -1.230167026 -0.905381156 -1.142569085 -0.908757221 -1.147756526
#>  [46] -1.001357004 -1.180938713 -1.813854339 -1.142433073 -1.146147630
#>  [51] -1.147757366 -0.739232990 -1.096344786 -1.354525240 -1.046292509
#>  [56] -1.063760042 -0.911734267 -1.034431184 -1.040308352 -1.037476707
#>  [61] -0.977631497 -1.008175595 -0.949178729 -1.026598749 -1.000895557
#>  [66] -1.069137807 -1.405770761 -1.282902684 -1.008919312 -0.864800853
#>  [71] -1.068741616 -0.742672477 -0.995370610 -1.085894111 -1.046580518
#>  [76] -0.964079417 -0.877483276 -1.027297991 -1.104032255 -0.962632053
#>  [81] -0.904232602 -0.877467100 -1.402177292 -1.287849684 -1.619600907
#>  [86] -0.885642638 -1.652767964 -0.985779007 -1.644007405 -1.166859783
#>  [91] -1.043143400 -0.984623463 -1.214100622 -1.619600907 -1.144407563
#>  [96] -1.304312620 -1.211414553 -1.001012201 -0.806740451 -1.354525240
#> [101] -0.966487426 -1.008175595 -0.990815294 -1.061374402 -0.891541468
#> [106] -1.053160336 -1.124484058 -0.971792888 -0.806026261 -0.775264091
#> [111] -1.205059148 -0.872039908 -0.892389628 -1.268263092 -0.943795908
#> [116] -1.083193698 -1.238196368 -0.855247083 -1.063760042 -1.055639844
#> [121] -1.007785582 -1.035674459 -1.205695553 -0.887938622 -0.902551425
#> [126] -1.278496322 -1.019787915 -1.208367111 -1.030931090 -1.069667183
#> [131] -0.907064484 -1.530875508 -1.071947413 -0.931809952 -1.026263858
#> [136] -1.526535992 -1.109655464 -0.948053611 -0.780694090 -0.916731822
#> [141] -0.854926934 -0.998240818 -1.033970809 -0.908757221 -0.903559523
#> [146] -0.819595879 -1.156713586 -0.903559523 -1.209688155 -0.876280045
#> [151] -1.068898550 -1.227733459 -1.046932966 -0.948962294 -1.075351218
#> [156] -1.103627985 -0.972030530 -0.995297134 -1.069555631 -1.340561970
#> [161] -1.037842544 -0.898928261 -1.033469430 -1.026112722 -1.218258656
#> [166] -0.964986443 -1.169445047 -1.089385659 -1.163999831 -1.123990563
#> [171] -1.027733262 -1.013896780 -0.890724433 -1.042093790 -0.911347375
#> [176] -0.914995620 -1.135854785 -0.970356887 -0.841869954 -1.008175595
#> [181] -0.855183045 -1.414911116 -1.210708876 -1.249107001 -1.128785873
#> [186] -1.138186307 -0.942236512 -1.100344973 -0.863185168 -1.013746575
#> [191] -0.869891760 -0.966133335 -0.874519436 -1.177841602 -1.119711826
#> [196] -1.148878341 -0.986079485 -0.746126043 -1.356353408 -0.735219304
#> [201] -1.105863546 -1.105084005 -0.922255465 -0.943635247 -1.015915700
#> [206] -0.753933441 -1.038071645 -1.047675180 -0.871513724 -0.981935009
#> [211] -0.901764987 -1.019183354 -0.996445993 -0.909857354 -1.008677783
#> [216] -1.230167026 -1.236755881 -1.307440598 -1.091945028 -1.291841745
#> [221] -1.223774578 -1.093963693 -1.138134058 -1.046350454 -1.026069705
#> [226] -1.167935778 -1.047675180 -0.900067469 -1.090504346 -1.243974248
#> [231] -1.002715189 -1.622588054 -1.041376508 -1.199507507 -1.205695553
#> [236] -1.116495473 -0.929456932 -1.370039302 -1.008175595 -1.145662664
#> [241] -0.886284792 -0.892785676 -1.126562810 -1.119558213 -1.085894111
#> [246] -1.037476707 -0.971311766 -1.144856000 -0.898525829 -0.892319859
#> [251] -1.469977834 -1.187206124 -0.948756668 -1.243103327 -0.935896722
#> [256] -0.918062721 -1.017939071 -1.076114273 -1.118002572 -1.202358600
#> [261] -1.100398788 -1.062018280 -1.008175595 -0.879037941 -0.924794731
#> [266] -1.217909165 -0.885907660 -1.942824284 -0.962225595 -1.124006860
#> [271] -0.983314725 -1.234329560 -1.005623136 -1.158573419 -1.176950708
#> [276] -0.428957818  0.012189863 -0.486951700  0.233442339 -0.195937232
#> [281] -0.080451711 -0.314301912 -0.078469904 -0.068359314 -0.515346080
#> [286]  0.027522347 -0.422747640 -0.547017943 -0.275295275 -0.686600179
#> [291] -0.529064099 -0.385978216 -0.368469550 -0.210950651 -0.166414946
#> [296] -0.337682148 -0.780554345  0.214327593 -0.247282604 -0.408385606
#> [301] -0.298109207 -0.259988674 -1.435464571 -0.163530444 -0.259587225
#> [306] -0.318695897 -0.316891109 -0.021768771 -0.483409839 -0.135617753
#> [311]  0.056634637 -0.582083772 -0.656086070 -0.525875983 -0.552459003
#> [316] -0.075587430 -0.676305597 -0.093704856 -0.116701874 -0.251852161
#> [321] -0.392050414 -0.549120485 -0.021671857 -0.467055159 -0.504663664
#> [326] -0.203145629  0.001250301 -3.203243171 -0.459827278 -0.214025071
#> [331] -0.326689539 -0.066975642 -0.569945761 -0.506537517 -0.138543861
#> [336] -0.132973022 -0.133277823 -0.164380679 -0.218658962 -0.066975642
#> [341] -0.011036599 -0.078762962 -0.180397856 -0.119167751 -0.024803795
#> [346] -0.143565777 -0.218526499 -0.578902854  0.033209387 -0.008817788
#> [351] -0.195272289 -0.343008139 -0.108169878 -0.131817637 -0.317774461
#> [356] -0.249381623  0.009996475 -0.078705432 -0.043191836 -0.337903826
#> [361] -0.678983723 -0.394767735 -0.532042677 -0.412270481  0.064681369
#> [366] -0.201828898  0.080302344  0.021983613  0.009899905 -0.273025138
#> [371] -0.009879200 -0.343077899 -0.008654474 -0.209960157 -0.403828767
#> [376] -0.165443021 -0.221395703 -0.426147368 -0.369369057 -0.553798226
#> [381] -0.886218188 -0.616719036 -0.098036692 -0.131787620 -0.089711764
#> [386] -0.408434778 -0.117855764 -0.601018047  0.022011724  0.094692459
#> [391] -0.444712813 -0.404969691 -0.032190646 -0.533679788 -0.426425805
#> [396]  0.229494989  0.025528617 -0.193993586 -0.298287852 -0.203714308
#> [401] -0.998858533 -0.318846660 -0.465223018 -0.150798611 -0.382835799
#> [406] -0.152336114  0.220389272 -0.934675232 -0.392121624 -0.112371143
#> [411]  0.006072157 -0.285898602 -0.110961265 -0.519771854  0.068535455
#> [416] -0.021768771 -0.166474664  0.108690515 -0.396023549 -0.678171298
#> [421] -0.591397383  0.172775595  0.045078655 -0.021579661 -0.022439870
#> [426] -0.366632061 -0.567698609 -0.348896178 -0.338459222  0.044135615
#> [431] -0.257630175 -0.217815950 -0.426425805 -0.604305406 -0.260659869
#> [436] -0.408843466 -0.217494774 -0.459783988 -0.309869972 -0.043119153
#> [441] -0.515904393 -0.477175296 -0.113681075 -0.552339040 -0.420222169
#> [446] -0.322945798 -0.153872572 -0.245367220 -0.449309414 -0.037912437
#> [451] -0.626065683 -0.032453939  0.045765724 -0.480040145 -1.623514835
#> [456] -0.111963334 -0.192595306 -0.361966378 -0.634193466 -0.093646960
#> [461]  0.061375155 -0.068755787 -0.105067295 -0.223505949 -0.365665405
#> [466] -0.212023280 -0.436867894 -0.182368905 -0.289800381 -0.242631685
#> [471] -1.035922137 -0.396584750 -0.160876553 -0.075034206 -0.166414946
#> [476] -0.223194271 -0.257009657 -0.172491681  0.009996475 -0.591397383
#> [481] -0.587245299 -0.587245299 -1.666292298 -0.328343846 -0.277103474
#> [486] -0.318695897 -0.373971893 -0.590806270 -0.252238915 -0.118865698
#> [491] -0.284484028  0.076587970 -0.241533949 -0.646625973 -0.197843716
#> [496] -0.311121357 -0.105225902 -0.240575270 -0.389098443 -0.016642511
#> [501] -0.432409159 -0.189644960 -0.174804921  0.048019974 -0.108307847
#> [506] -0.314301912 -0.428107912 -0.879031175 -0.482207766 -0.155950475
#> [511] -0.336749035 -0.718994764 -0.359657666 -0.657175398 -0.007180822
#> [516]  0.020298891 -0.043448886 -0.062859657 -0.323759058 -0.303421045
#> [521] -0.358092107  0.186266492  0.078865651  0.519517807  0.269921879
#> [526]  0.597018315  0.514982565  0.627700759  0.696775216  0.669326734
#> [531]  0.457606862  0.681160400  0.519036164  0.613267542  0.222525068
#> [536]  0.397956889  0.133771273  0.370296434  0.563772499  0.594610656
#> [541] -0.336323633  0.736671476  0.474222608 -0.777125948  0.203868258
#> [546] -0.753680313 -1.444006915  0.028944380  0.202000659  0.504085790
#> [551]  0.600514364  0.697516671  0.478017443  0.289020489  0.303096324
#> [556]  0.737926651  0.187946343  0.586031948  0.425896367  0.224086646
#> [561] -0.215965800  0.402547072  0.398890152  0.593142760  0.315282144
#> [566]  0.423323409  0.332137510  0.662126525 -0.442890861  0.140693026
#> [571]  0.342805540  0.430517086  0.358493991  0.083998501  0.296390711
#> [576]  0.437550189  0.178746020  0.382561031  0.879850603  0.640128186
#> [581]  0.631271930  0.172684020  0.328363535  0.340478087  0.344914356
#> [586]  0.461381123  0.336833527  0.448599069  0.392974516 -0.460327789
#> [591]  0.445830685  0.699633102  0.365061432  0.446151309  0.102516805
#> [596]  0.499378446  0.740927085  0.239177056  0.591873666 -0.309058152
#> [601]  0.282965584  0.733434678  0.657550625  0.354833012  0.662908850
#> [606] -0.511091884  0.303787760  0.562003625  0.164701428  0.416301559
#> [611] -0.386647626  0.251650522  0.525854865  0.388067761  0.606624948
#> [616]  0.539410510  0.299480721 -0.346216422  0.325213100  0.245366654
#> [621]  0.112192006  0.335399827  0.184816359  0.587883132 -0.056364660
#> [626]  0.166293553  0.890644865  0.284874655 -0.051583362 -0.033817304
#> [631]  0.143137537  0.662689484  0.603571152  0.605066194  0.732388316
#> [636]  0.192572699 -0.144711816  0.581012587  0.642500012  0.134367262
#> [641]  0.551083996  0.151382964 -0.253267219  0.438174389 -0.115021445
#> [646]  0.586168765  0.536178304  0.316355287  0.418002000  0.452139782
#> [651]  0.621197281 -0.033334905  0.599417705  0.542257544  0.713252044
#> [656]  0.242526676  0.501978521  0.262166414 -0.005694922  0.454746475
#> [661]  0.513719465  0.316235408  0.247407055  0.663438961  0.692741990
#> [666] -0.115021445  0.347554874  0.679070171 -1.605504577  0.251727555
#> [671]  0.439592865  0.653896599  0.633110036  0.464877821  0.201002880
#> [676]  0.585373477  0.045002166 -0.491437798  0.338167713  0.325213100
#> [681] -0.412061553 -0.483848856  0.492190273  0.336145050  0.490407748
#> [686]  0.389378315  0.470698725  0.755551737  0.908657687  0.214232189
#> [691]  0.394056153  0.664081231  0.068285188 -0.171097400  0.547419585
#> [696]  0.691898486  0.571540486  0.340873165  0.218322588  1.002757052
#> [701]  1.356677961  0.834785710  0.354696859  0.744029931  0.715925297
#> [706]  1.309592504  0.992691881  0.728488364  0.055196779  0.580749847
#> [711]  1.267965730  1.106234686 -0.529716606  0.646177569  0.728488364
#> [716]  0.786831056  1.200323736  1.056427539  0.975485893  1.063048750
#> [721]  0.724519298  0.584237501  1.341089279  0.988315954  0.571606438
#> [726]  0.548187113  0.962956129  1.089583709  0.864715587  0.680253395
#> [731]  0.996219570  1.284018492  1.033070375  0.648937838  0.879841375
#> [736]  0.195576864  1.159139234  1.258800743  1.204909638  0.845643348
#> [741]  0.498697782 -0.307442266  0.858101141  0.047578672  0.518734644
#> [746]  1.235836551  0.513336649  0.830809006  1.177904806 -0.033496744
#> [751]  0.482656190  1.384674734  1.014770415  0.381999433  1.063977290
#> [756]  0.910101338  1.263677852  0.842091706  0.920701172  0.980049549
#> [761]  0.893306066  0.860981680  0.750464266  1.312229578  0.809367053
#> [766]  0.430300600  0.800981208  0.265901485  0.693667409  1.305466879
#> [771] -0.571703071  0.719445957  1.304858816  1.029244357  0.639128749
#> [776]  0.732017629  0.261334067  1.253998197  1.147305810  0.986689241
#> [781]  1.484574671  1.209553031  1.336680200  1.291670360  1.206834988
#> [786]  0.631739763  1.421402377  1.141184812  1.478345008  1.875185731
#> [791]  1.545872780  1.928089634  1.373364059  1.703288232 -0.202984254
#> [796]  1.483979556  1.590249896  0.938605407 -0.641951950  1.588421380
#> [801]  1.890869042  1.471131865 -0.750468504  1.363750023  1.484561925
#> [806]  1.316949737  1.433373778  1.130759666  1.546946673  1.032029891
#> [811]  1.668109392 -0.101105536  1.847011914  1.868673425  0.288952542
#> [816]  1.083850249  1.314937465  1.647356614  1.609967994  1.496324671
#> [821]  1.245133194  0.602223990  1.304139103  1.649362248  1.295280146
#> [826]  1.766545874  1.604903216  1.701468207  1.712004460  1.000031929
#> [831]  1.429568676  0.998350240  1.835700114  1.603807462  0.660100969
#> [836]  1.527711874  1.528806779  1.853279336  1.215533445  1.237601607
#> [841]  0.846098431  1.621856210  1.664702555  1.642921940  1.846403544
#> [846]  1.426893831  1.280521753  1.344358498  1.759744675  1.166530892
#> [851]  1.746174128  2.055465244  1.974170248  2.063123947  1.245272883
#> [856]  1.742051459  1.518886151  1.808431542  2.023760660  1.336292355
#> [861]  1.478206681  2.099350713  1.475595533  0.757831536  2.101226703
#> [866]  1.748557330  2.481343156  1.841133110  2.224487173  1.434510973
#> [871]  2.260543070 -0.019413184  1.929503607  1.620977811  1.875863863
#> [876]  2.301469984  2.383206111  2.424852003  1.395248178  1.971667992
#> [881]  2.104440726  1.876418502  2.620786967  2.266128895  2.004538978
#> [886]  2.174582558  2.568886980  2.138417721  2.692261868  2.462677207
#> [891]  2.469038244  2.729548699  2.561618729  2.791290378  2.875341226
#> [896]  2.944757758  2.382379970  3.213573100  3.161991196  2.566325583
#> [901]  1.332515691  2.564123905  2.408105375  2.408954734  2.132311738
#> [906]  2.419134404  2.929986547  3.141813307  1.956009463  3.812992438
#> [911]  4.919918464  3.632313931  5.091665444  5.355691031  4.964986453
#>